### 2021-06-10 #3350 @whedon
[PRE REVIEW]: TsgFEM: Tensegrity Finite Element Method
**Submitting author:** @Muhao-Chen (<a href="http://orcid.org/0000-0003-1812-6835">Muhao Chen</a>)
**Repository:** <a href="https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM" target ="_blank">https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM</a>
**Version:** v1.1
