### 2021-06-10 #3350 @Kevin-Mattheus-Moerman
@Muhao-Chen thanks for this submission to JOSS. We have some questions in relation to this work.

This looks like it is a continuation of work you and your co-workers previously submitted to JOSS, namely MOTES ([paper](https://joss.theoj.org/papers/10.21105/joss.01613), [review](https://github.com/openjournals/joss-reviews/issues/1613), [software repository](https://github.com/ramaniitrgoyal92/Modeling_of_Tensegrity_Structures_MOTES)).  Indeed some functionality (e.g. [tenseg_plot.m](https://github.com/ramaniitrgoyal92/Modeling_of_Tensegrity_Structures_MOTES/blob/master/Function_Library/tenseg_plot.m) in MOTES and [tenseg_plot.m](https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/Function_library/tenseg_plot.m) in TsgFEM) seems fully duplicated. Also the README may have been copied from MOTES as the [set-up section for TsgFEM](https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#setupm) refers to MOTES.

Can you clarify how the current new submission, TsgFEM, differs from MOTES? Can you clearly make the case why TsgFEM requires an independent publication on top of, and distinct from MOTES? Can you please also comment on whether it would not be better to integrate TsgFEM+MOTES into a single library and to consider publishing that?

Can you clarify the timeline of development for this work? Based on GitHub uploads it looks like development just started,  but [the user guide](https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/User_Guide/User_Guide_Tensegrity_Finite_Element_Method_(TsgFEM)%20.pdf) appears to refer to development from 2019.
