### 2021-06-10 #3350 @Kevin-Mattheus-Moerman
@Muhao-Chen can you check if all those references are actually cited in the paper? :point_up: Perhaps clean up the .bib file so it is only the papers from the paper and also add those missing DOIs please. You can call `@whedon generate pdf` to update the paper.
