+++ JOSS_paper/paper.md
@@ -1,8 +1,5 @@
 ---
-
 title: 'TsgFEM: Tensegrity Finite Element Method'
-
-
 tags:
   - Tensegrity systems
   - Multibody dynamics
@@ -11,33 +8,26 @@ tags:
   - Finite element method
   - Linearized tensegrity dynamics
   - Elastic and plastic deformation
-
 authors:
   - name: Shuo Ma
     orcid: 0000-0003-3789-2893
     affiliation: 1
-
   - name: Muhao Chen^[corresponding author]
     orcid: 0000-0003-1812-6835
     affiliation: 2
-
   - name: Robert E. Skelton
     orcid: 0000-0001-6503-9115
     affiliation: 2
-    
 affiliations:
  - name: College of Civil Engineering, Zhejiang University of Technology, Hanzhou, Zhejiang, China
    index: 1
-
  - name: Department of Aerospace Engineering, Texas A&M University, College Station, Texas, USA
    index: 2
-
 date: 4 June 2021
 bibliography: paper.bib
 ---
 
 # Summary
-
 
 Tensegrity is a coined word by Buckminister Fuller [@fuller1982synergetics] for the art form created by Ioganson (1921) and Snelson (1948) [@lalvani1996origins] to represent a stable network of compressive members (bars/struts) and tensile members (strings/cables) [@chen2020general]. From the definition, it is straightforward to see that the fundamental property of the tensegrity is that all the bars and strings are axially loaded. Since the bars and strings are best in taking compression and tension and there is no material bending, the structure mass can be greatly reduced. Indeed, biological structures also indicate that tensegrity concepts yield the most efficient structures. For example, the bones, muscles, and elbows of animals and humans are tensegrity models. Wang et al. found that microtubules and microfilaments in the living cells work as compressive and tensile members to change the traction of the cell surfaces [@wang2001mechanical]. Simmons et al. showed that the DNA bundles are consistent with tensegrity prism [@liedl2010self]. After decades of study, many lightweight structures have been redesigned by the tensegrity paradigm. For example, Skelton and de Oliveira proved that T-Bar and D-Bar structures require less mass than a signal rod in taking compressive buckling load [@skelton2009tensegrity]. Ma et al. showed a mass efficient tensegrity cantilever structure subject to yield and buckling constraints [@ma2020design]. Barbarigos et al. designed and analyzed a lightweight pedestrian bridge [@rhode2010designing]. And many space applications are employing tensegrity solutions, for example, lightweight space habitat [@chen2021review], deployable lunar tower for space mining [@chen2020deployable], and planetary landers [@luo2017analysis;@sabelhaus2015system]. Moreover, the many advantages of tensegrity have also attracted researchers to find new ways to design soft robotics, i.e., six-bar tensegrity robot [@booth2020surface;@wang2020first], robotic spine [@sabelhaus2020model], morphing wings [@chen2020design], robotic fish [@chen2019swimming], debris capturing robot [@feng2021design].
 
@@ -49,10 +39,7 @@ The name of this software, TsgFEM, is suggested to be pronounced as Tenseg FEM. 
 
[…]
