+++ JOSS_paper/paper.bib
@@ -1,49 +1,25 @@
-@article{jensen2007finite,
-  title={Finite element analysis of tensegrity structures in offshore aquaculture installations},
-  author={Jensen, {\O}sten and Wroldsen, Anders Sunde and Lader, P{\aa}l Furset and Fredheim, Arne and Heide, Mats},
-  journal={Aquacultural Engineering},
-  volume={36},
-  number={3},
-  pages={272--284},
-  year={2007},
-  publisher={Elsevier}
+@book{fuller1982synergetics,
+  title={Synergetics: explorations in the geometry of thinking},
+  author={Fuller, R Buckminster},
+  year={1982},
+  publisher={Estate of R. Buckminster Fuller}
 }
-
-@article{kan2018nonlinear,
-  title={Nonlinear dynamic and deployment analysis of clustered tensegrity structures using a positional formulation FEM},
-  author={Kan, Ziyun and Peng, Haijun and Chen, Biaoshong and Zhong, Wanxie},
+@article{lalvani1996origins,
+  title={Origins of tensegrity: views of Emmerich, Fuller and Snelson},
+  author={Lalvani, Haresh},
+  journal={International Journal of Space Structures},
+  volume={11},
+  number={1-2},
+  pages={27--27},
+  year={1996},
+  publisher={SAGE Publications Sage UK: London, England}
+}
+@article{chen2020general,
+  title={A general approach to minimal mass tensegrity},
+  author={Chen, Muhao and Skelton, Robert E},
   journal={Composite Structures},
-  volume={187},
-  pages={241--258},
-  year={2018},
-  publisher={Elsevier}
-}
-
-@inproceedings{zgoul2012static,
-  title={Static Analysis of a Tensegrity Bridge using the Finite Element Method},
-  author={Zgoul, MH and Alzamer, A and Elayyan, MR and Quran, MA},
-  booktitle={International Conference on Applications and Design in Mechanical Engineering},
-  pages={27--28},
-  year={2012}
-}
-
-@article{rhode2010designing,
-  title={Designing tensegrity modules for pedestrian bridges},
-  author={Rhode-Barbarigos, Landolf and Ali, Nizar Bel Hadj and Motro, Ren{\'e} and Smith, Ian FC},
-  journal={Engineering Structures},
-  volume={32},
-  number={4},
-  pages={1158--1167},
-  year={2010},
-  publisher={Elsevier}
-}
-@article{wang2021minimal,
-  title={Minimal mass design of active tensegrity structures},
-  author={Wang, Yafeng and Xu, Xian and Luo, Yaozhi},
-  journal={Engineering Structures},
-  volume={234},
-  pages={111965},
-  year={2021},
+  pages={112454},
+  year={2020},
   publisher={Elsevier}
 }
 @article{wang2001mechanical,
@@ -56,7 +32,6 @@
   year={2001},
   publisher={National Acad Sciences}
 }
-
 @article{liedl2010self,
   title={Self-assembly of three-dimensional prestressed tensegrity structures from DNA},
   author={Liedl, Tim and H{\"o}gberg, Bj{\"o}rn and Tytell, Jessica and Ingber, Donald E and Shih, William M},
@@ -67,124 +42,13 @@
   year={2010},
   publisher={Nature Publishing Group}
 }
-
-@article{simmons1996molecular,
-  title={Molecular orientation and two-component nature of the crystalline fraction of spider dragline silk},
-  author={Simmons, Alexandra H and Michal, Carl A and Jelinski, Lynn W},
-  journal={Science},
-  volume={271},
-  number={5245},
-  pages={84--87},
-  year={1996},
[…]
