+++ JOSS_paper/paper.bib
@@ -2,7 +2,8 @@
   title={Synergetics: explorations in the geometry of thinking},
   author={Fuller, R Buckminster},
   year={1982},
-  publisher={Estate of R. Buckminster Fuller}
+  publisher={Estate of R. Buckminster Fuller},
+  doi={10.2307/3103256}
 }
 @article{lalvani1996origins,
   title={Origins of tensegrity: views of Emmerich, Fuller and Snelson},
@@ -12,7 +13,8 @@
   number={1-2},
   pages={27--27},
   year={1996},
-  publisher={SAGE Publications Sage UK: London, England}
+  publisher={SAGE Publications Sage UK: London, England},
+  doi={10.1177/026635119601-204}
 }
 @article{chen2020general,
   title={A general approach to minimal mass tensegrity},
@@ -20,7 +22,8 @@
   journal={Composite Structures},
   pages={112454},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.compstruct.2020.112454}
 }
 @article{wang2001mechanical,
   title={Mechanical behavior in living cells consistent with the tensegrity model},
@@ -30,7 +33,8 @@
   number={14},
   pages={7765--7770},
   year={2001},
-  publisher={National Acad Sciences}
+  publisher={National Acad Sciences},
+  doi={10.1073/pnas.141199598}
 }
 @article{liedl2010self,
   title={Self-assembly of three-dimensional prestressed tensegrity structures from DNA},
@@ -40,14 +44,16 @@
   number={7},
   pages={520--524},
   year={2010},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/nnano.2010.107}
 }
 @book{skelton2009tensegrity,
   title={Tensegrity systems},
   author={Skelton, Robert E and de Oliveira, Mauricio C},
   volume={1},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-0-387-74242-7}
 }
 @article{ma2020design,
   title={Design of a new tensegrity cantilever structure},
@@ -55,7 +61,8 @@
   journal={Composite Structures},
   pages={112188},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.compstruct.2020.112188}
 }
 @article{rhode2010designing,
   title={Designing tensegrity modules for pedestrian bridges},
@@ -65,7 +72,8 @@
   number={4},
   pages={1158--1167},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.engstruct.2009.12.042}
 }
 @article{chen2021review,
   title={Review of space habitat designs for long term space explorations},
@@ -74,13 +82,15 @@
   volume={122},
   pages={100692},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.paerosci.2020.100692}
 }
 @article{chen2020deployable,
   title={Deployable Tensegrity Lunar Tower},
   author={Chen, Muhao and Goyal, Raman and Majji, Manoranjan and Skelton, Robert E},
   journal={Earth and Space 2021},
-  year={2021}
+  year={2021},
+  doi={10.1061/9780784483374.100}
 }
 @article{luo2017analysis,
   title={Analysis for feasibility of the method for bars driving the ball tensegrity robot},
@@ -89,7 +99,8 @@
   volume={9},
   number={5},
   year={2017},
-  publisher={American Society of Mechanical Engineers Digital Collection}
[…]
