### 2021-06-10 #3350 @Muhao-Chen
@Kevin-Mattheus-Moerman Thank you so much for your help! I removed all the unused bib files and also added the missing DOIs. The pdf format looks good now. Please let me know if there are any questions. Thank you!
