### 2021-06-21 #3350 @Kevin-Mattheus-Moerman
@diehlpk thanks for handling this. FYI here are two other papers I was involved in which feature tensegrity modelling. Perhaps some of the authors/reviewers for these can help review too:

- https://github.com/openjournals/joss-reviews/issues/1042
- https://github.com/openjournals/joss-reviews/issues/1613 (note this one shares authors with the current submission, see also my comments :point_up: )
