### 2021-06-21 #3390 @whedon
[REVIEW]: TsgFEM: Tensegrity Finite Element Method
**Submitting author:** @Muhao-Chen (<a href="http://orcid.org/0000-0003-1812-6835">Muhao Chen</a>)
**Repository:** https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM
**Branch with paper.md** (empty if default branch):
**Version:** v1.2
