### 2021-07-31 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#1 @HaoZeke
[JOSS] Automated Tests
Part of https://github.com/openjournals/joss-reviews/issues/3390.

It is necessary to include [unit tests](https://www.mathworks.com/help/matlab/matlab-unit-test-framework.html) and instructions to run them; without which over time the maintainability of the software will suffer.
