### 2021-07-31 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#2 @HaoZeke
[JOSS] Paper updates
Part of https://github.com/openjournals/joss-reviews/issues/3390.

The software paper should:
- include some of the algorithms and a clearer methodology section describing the software suite
- describe the structure and layout of the code
- provide a graphical overview of accuracy v.s. size
- describe the examples (since they are mentioned for the accuracy)
- describe the interface to ANSYS in more detail
