### 2021-08-01 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#2 @likask
My suggestions in addition to @HaoZeke

- Explain the name and purpose of the code in the first paragraph. Consider breaking summary—introduction with literature review and physical and numerical assumptions.
- Do not mix assumptions about the physical problem and numerical discretisation.
- Consider adding figures with results and deformation on the first page. That will give short insight into what your code does.
-  Do not use "Accurate FEM dynamics". You have quasi rigid members for large stiffness and elastic or elasto-plastic membres.  FEM is discretisation method.
- You can mention solution schemes, which you are using to solve the nonlinear problem, dynamic problem, or natural frequency problem.
- Using terms like "natural frequency errors are about 10−14" is imprecise; we do not know if this is absolute or relative error. I would suggest writing that both codes, Ansys and TsgFEM give errors which in numerical precision for some particular tested problems.
