### 2021-08-01 #3390 @diehlpk
> OK, the reviewer has been re-invited.
>
> @likask please accept the invite by clicking this link: https://github.com/openjournals/joss-reviews/invitations

@likask Please accept the invitation again und you should be able to check the check boxes.
