### 2021-08-01 #3390 @likask
Software is easy to install; authors provide a simple Matlab script that setup all paths necessary to run code. The user guide provides with the code gives a straightforward explanation of how to run code and how to analyze results. Examples are a good start to create own structures.

The physical problem itself is classical, and standard finite element techniques can be applied. The strength of the  TsgFEM is its simplicity, compactness, and simplified installation in the Matlab environment.

I add some issues to the code repository. All related to JOSS paper. Once addressed, I am happy to accept.
