### 2021-08-23 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
JOSS review general comments
Below are my review comments for https://github.com/openjournals/joss-reviews/issues/3390

- [ ] In `setup.m' the function library path is not added properly due to use of capital L (an issue on Linux), i.e. it should read:

```
addpath( genpath('Function_library') );
```

- [ ] Check the names of functions/examples e.g. `main_tower_dyanmic_simple.m` should likely be renamed to `main_tower_dynamic_simple.m`
- [ ] The use of `clear all;` has been simplified to `clear;` in modern MATLAB version
- [ ] Consider reviewing the code for bad MATLAB style e.g. not allocating arrays and having them grow in for loops. Further more where possible for-loops should be avoided to speed up performance.
Example of a for-loop without allocation (from `main_tower_dyanmic_simple.m`), here of `N`, MATLAB flags this with a warning too:
```matlab
for i=1:p               % nodal coordinate matrix N
    N(:,i)=R*[cos(2*pi*(i-1)/p),sin(2*pi*(i-1)/p),0];
end
```
It is better practice to use something like:
```matlab
N=zeros(p,3); %Allocate N array
for i=1:p               % nodal coordinate matrix N
    N(:,i)=R*[cos(2*pi*(i-1)/p),sin(2*pi*(i-1)/p),0];
end
```
Furthermore upon inspection the loop can be avoided to speed up the code since the array can be create "in one go":
```matlab
t=2*pi*((1:p)-1)./p;
N=R*[cos(t); sin(t); zeros(1,p)];
```

- [ ] Another example of code style (which MATLAB also flags with warnings) is the use of `[1:3]` instead of the proper `(1:3)`
- [ ] It is recommended to use code "indenting", a simple way to do this is to execute CTRL-A (select all) followed by CTRL-I (indent all), i.e. this would transform:
 ```matlab
if auto_dt
dt=pi/(8*max(omega)); 	% time step dt is 1/8 of the smallest period, guarantee convergence in solving ODE
end
```
to
```matlab
if auto_dt
    dt=pi/(8*max(omega)); 	% time step dt is 1/8 of the smallest period, guarantee convergence in solving ODE
end
```

- [ ] Is the use of global variables strictly speaking needed (e.g. in `ode4_tsg.m`)? Performance can likely be improved if they can be avoided.
- [ ] The function `ode4_tsg.m` features `disp(ti);` which produces `ti` on the command window. Without further information this is just a number that streams on the command window (and slows performance). If you really like to display this information I suggest/recommend that you display a comment on what this variable is too. E.g. if you labelled this as `Current time: ` one could use (note also I used `sprintf` with the format `'%.5f'` to directly specify the output format and number of decimal places):
```matlab
disp(['Current time: ',sprintf('%.5f',pi)]);
```
Also I suggest/recommend adding an option to not output this number (e.g. implement something like "silentMode"). Note this is just a suggestion that I would look for if I used this code.
If the intention is to provide an indication of "completion", then another suggestion/improvement would be to actually compute the % completion and to show that, and one could use `waitbar` rather than dumping content in the command window.
- [ ] In the function `tenseg_video.m` the function `tenseg_plot.m` is called. The latter appears to be designed to receive a figure handle as one of its inputs. However, within `tenseg_video.m` it simply passes the number `99`, instead of a figure handle.
```matlab
tenseg_plot(N,C_b,C_s,99,[],[],[],R3Ddata);
```
It seems the authors assume the user does not use more than 98 figures and that 99 should be a new figure (In this demo this creates figures 1 up to 10 and a new figure 99). Note however if figure 99 exists this code would override. It would be better practice (and less confusing) to simply open a new empty figure and to pass the figure handle properly. E.g. it looks like the following would do the trick (if one has 10 figures already open this would create an figure numbered 11 rather than 99):
```matlab
hf=figure;
tenseg_plot(N,C_b,C_s,hf,[],[],[],R3Ddata);
```

- [ ] I get the following error upon running `main_10seg_truss_simple.m`
```error
Error using wgifc
Can only append to GIF89a format GIFs.

Error in writegif (line 306)
wgifc(mat, map,
filename,writemode,disposalmethod,delaytime,...

Error in imwrite (line 566)
        feval(fmt_s.write, data, map, filename,
        paramPairs{:});

Error in tenseg_savegif_forever (line 49)
    imwrite(f,map,filename,'writemode','append','delaytime',dt);

Error in tenseg_video (line 43)
        tenseg_savegif_forever(name);

Error in main_10seg_truss_simple (line 172)
tenseg_video(n_t,C_b,C_s,[0,10.5,-1.5,1.2,0-1,1],100,name,savevideo);
```

- [ ] In `Main_Double_Pendulum_Analytical.m` the parameter `'RelTol'` is initiated as `1e-16`, however I immediately get a MATLAB warning stating: `Warning: RelTol has been increased to 2.22045e-14`. Please check the code and that all parameters (across all example) are indeed suitable.
- [ ] The files to run for the dynamic examples feature both `main_...` and `Main_...`, please choose one type and stick with this consistently. I also suggest renaming the static examples which currently start with `example_...`, perhaps use either `main_...` or `example_...` for all example files.
- [ ] In `Example_folding_3d_Dbar.m` (and several other examples) have an odd comment `%ºÉÔØ×Ó²½`, is this intentional?
- [ ] I get the same gif related error described above in: `Example_lander.m` (line 157)
- [ ] In `main_Jasen_linkage_2.m` (and several other static examples) several figure handles show up on the command window as `ans=...`, in other words perhaps a semi-colon `;` is missing to suppress these.
- [ ] For most examples it seems the majority of time is spent writing the visualisation videos/gifs, perhaps consider having the saving be optional so the demos run fast and could skip this export.
- [ ] In my personal project [GIBBON](https://github.com/gibbonCode/GIBBON) I use a MATLAB script called 'testGibbon' (https://github.com/gibbonCode/GIBBON/blob/master/lib/testGibbon.m) to run all tests/demos automatically. Consider adopting a similar automated testing scheme. Running all automatically (or even manually) would be simplified if all examples could be found in a single folder but this is not required.
