### 2021-08-23 #3390 @Kevin-Mattheus-Moerman
@diehlpk thanks for the reminders, apologies these announcement were snowed under for a while. I started review and my initial comments are here:
https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/issues/3
