### 2021-08-23 #3390 @Kevin-Mattheus-Moerman
@Muhao-Chen thanks for submitting this interesting work. I have downloaded and tested the software and I enjoyed running the various examples. I think this is in principle a nice submission but I do think the work can benefit from a lot of minor changes which I highlight in this issue:
https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/issues/3
One major point is automated testing. I provide some suggestions on this in the above issue too.

I have not yet read the paper but will do so shortly and may come back with more comments.
