### 2021-09-04 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
@Muhao-Chen let me know if you have any questions. Happy to help explain more.
