+++ Software_Verification_and_Examples/Statics_Examples/05_cable_dome/cable_dome.txt
@@ -4,7 +4,7 @@ finish
 /clear 
 /filename,tower  
 /title,the analysis of tower  
-!Unit:mï¼ŒNï¼ŒPaï¼Œs
+!Unit:m£¬N£¬Pa£¬s
 
 /prep7
 !specify element type 
@@ -761,7 +761,7 @@ latt,1,,1,,,,9  !specify section area
 LMESH,ALL
 finish
 
-!First solve for self-equilibriumï¼‰
+!First solve for self-equilibrium£©
 /SOLU
 ANTYPE,0 
 NLGEO!consider large deformation 
