### 2021-10-02 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
@Muhao-Chen are you able to address these issues?
