### 2021-10-29 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @diehlpk
@Muhao-Chen Please start working on these remarks, so we can proceed with the review.
