+++ Software_Verification_and_Examples/Statics_Examples/05_cable_dome/main_cable_dome.m
@@ -8,8 +8,7 @@
 clc;clearvars;close all;
 % global l  Eb Es
 
-%EXAMPLE
-clc; clear all; close all;
+
 % Global variable
 [consti_data,Eb,Es,sigmab,sigmas,rho_b,rho_s]=material_lib('Steel_Q345','Steel_string');
 material{1}='linear_elastic'; % index for material properties:'linear_elastic' multielastic plastic
