### 2021-10-30 db22dbd
display current time
