+++ Software_Verification_and_Examples/Statics_Examples/05_cable_dome/README.md
@@ -1,1 +1,1 @@
-This folder has one M file, 'main_cable_dome.m', conducting pseudo-static analysis of a Levy cable dome in vertical load exerting on every top node.
+This folder has one M file, 'Main_cable_dome.m', conducting pseudo-static analysis of a Levy cable dome in vertical load exerting on every top node.
