+++ Software_Verification_and_Examples/Statics_Examples/01_tower/README.md
@@ -1,1 +1,1 @@
-This folder has one M file, 'Example_tower_static.m', conducting the static analysis of a two-layer tensegrity prism in compression and tension force on the three top nodes.
+This folder has one M file, 'Main_tower_static.m', conducting the static analysis of a two-layer tensegrity prism in compression and tension force on the three top nodes.
