+++ Software_Verification_and_Examples/Statics_Examples/03_lander/README.md
@@ -1,1 +1,1 @@
-This folder has one M file, 'Example_lander.m', conducting a pseudo-static analysis of an expanded octahedron tensegrity with the rest length of one string changing.
+This folder has one M file, 'Mian_lander.m', conducting a pseudo-static analysis of an expanded octahedron tensegrity with the rest length of one string changing.
