+++ Software_Verification_and_Examples/Dynamics_Examples/02_cantilever_truss/main_10seg_truss_simple.m
@@ -36,8 +36,8 @@ tf=2;                   % final time of dynamic simulation
 out_dt=1e-2;            % output data interval(approximately, not exatly)
 lumped=0;    % use lumped matrix 1-yes,0-no
 saveimg=0;              % save image or not (1) yes (0)no
-savedata=1;             % save data or not (1) yes (0)no
-savevideo=1;            % make video(1) or not(0)
+savedata=0;             % save data or not (1) yes (0)no
+savevideo=0;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
 % move_ground=0;          % for earthquake, use pinned nodes motion(1) or add inertia force in free node(0) 
 %% N C of the structure
