+++ Software_Verification_and_Examples/Dynamics_Examples/01_tower/main_tower_dynamic_simple.m
@@ -33,9 +33,9 @@ auto_dt=0;              % use(1 or 0) auto time step, converengency is guarantee
 tf=2;                   % final time of dynamic simulation
 out_dt=0.02;            % output data interval(approximately, not exatly)
 lumped=0;               % use lumped matrix 1-yes,0-no
-saveimg=1;              % save image or not (1) yes (0)no
-savedata=1;             % save data or not (1) yes (0)no
-savevideo=1;            % make video(1) or not(0)
+saveimg=0;              % save image or not (1) yes (0)no
+savedata=0;             % save data or not (1) yes (0)no
+savevideo=0;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
 % move_ground=0;          % for earthquake, use pinned nodes motion(1) or add inertia force in free node(0)
 
