+++ Software_Verification_and_Examples/Statics_Examples/04_Jasen_linkage/example1/main_Jasen_linkage_1.m
@@ -24,8 +24,8 @@ c_s=0.1;           % coefficient of safty of strings 0.3
 substep=100;                                     %ºÉÔØ×Ó²½
 lumped=0;               % use lumped matrix 1-yes,0-no
 saveimg=0;              % save image or not (1) yes (0)no
-savedata=1;             % save data or not (1) yes (0)no
-savevideo=1;            % make video(1) or not(0)
+savedata=0;             % save data or not (1) yes (0)no
+savevideo=0;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
 
 %% N C of the structure
