+++ Software_Verification_and_Examples/Statics_Examples/01_tower/Main_tower_static.m
@@ -32,9 +32,9 @@ c_s=0.1;           % coefficient of safty of strings 0.3
 
 substep=100;                                     %ºÉÔØ×Ó²½
 lumped=0;               % use lumped matrix 1-yes,0-no
-saveimg=1;              % save image or not (1) yes (0)no
-savedata=1;             % save data or not (1) yes (0)no
-savevideo=1;            % make video(1) or not(0)
+saveimg=0;              % save image or not (1) yes (0)no
+savedata=0;             % save data or not (1) yes (0)no
+savevideo=0;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
 
 %% N C of the structure
