+++ Software_Verification_and_Examples/Statics_Examples/04_Jasen_linkage/example2/main_Jasen_linkage_2.m
@@ -3,7 +3,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
-% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
+% * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 clc;clearvars;close all;
 % global l  Eb Es
@@ -32,17 +32,17 @@ gravity=0;              % consider gravity 1 for yes, 0 for no
 % Manually specify node positions (accurate).
 N0=[   38.0000   38.0000   -8.7357  -39.6678         0  -19.4476   17.0047   30.3109
     7.8000   22.8000   40.5702   -5.8717         0  -39.6874  -35.4306  -82.5894
-         0         0         0         0         0         0         0         0];
-     N0(1,:)=N0(1,:)-38;
-     N1=N0;N1(3,:)=-5*ones(1,8);
-      N2=N0;N2(3,:)=-10*ones(1,8);
-      N=[N0,N1,N2];
+    0         0         0         0         0         0         0         0];
+N0(1,:)=N0(1,:)-38;
+N1=N0;N1(3,:)=-5*ones(1,8);
+N2=N0;N2(3,:)=-10*ones(1,8);
+N=[N0,N1,N2];
 % N=kron([1 1 1],N0);
 
 % Manually specify connectivity indices.
 C_s_in=[];  % This is indicating that string connection
 C_b_in0 = [1 2;2 3;3 4;3 5;4 5;4 6;5 7;6 7;6 8;7 8;2 7];  % Similarly, this is saying bar 1 connects node 1 to node 2,
-C_b_in=[C_b_in0; C_b_in0+8; C_b_in0+16];   
+C_b_in=[C_b_in0; C_b_in0+8; C_b_in0+16];
 % Convert the above matrices into full connectivity matrices.
 C_b = tenseg_ind2C(C_b_in,N);
 C_s = tenseg_ind2C(C_s_in,N);
@@ -58,9 +58,9 @@ pinned_X=([1 5 9 13 17 21])'; pinned_Y=([1 5 9 13 17 21])'; pinned_Z=(1:nn)';
 
 %% generate group index for tensegrity torus structure
 gr=[];
-Gp=tenseg_str_gp(gr,C);    %generate group matrix 
+Gp=tenseg_str_gp(gr,C);    %generate group matrix
 
-%% self-stress design 
+%% self-stress design
 %Calculate equilibrium matrix and member length
 [A_1a,A_1ag,A_2a,A_2ag,l,l_gp]=tenseg_equilibrium_matrix1(N,C,Gp,Ia);
 
@@ -71,8 +71,8 @@ Gp=tenseg_str_gp(gr,C);    %generate group matrix
 w0=zeros(numel(N),1); w0a=Ia'*w0;
 
 %prestress design
-index_gp=[1];                 % number of groups with designed force
-fd=[0];              % force in bar is given as -1000
+index_gp=1;                 % number of groups with designed force
+fd=0;              % force in bar is given as -1000
 [q_gp,t_gp,q,t]=tenseg_prestress_design(Gp,l,l_gp,A_1ag,V2,w0a,index_gp,fd);    %prestress design
 
 %% cross sectional design
@@ -104,7 +104,7 @@ num_plt=1:2;        % number of modes to plot
[…]
