### 2021-11-08 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#2 @diehlpk
@Muhao-Chen Could you please have a look here as well?
