### 2021-11-24 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @diehlpk
@Kevin-Mattheus-Moerman It seems that @Muhao-Chen has addressed some of your comments. Can you please have a look if he resolved the issues?
