### 2021-11-24 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#1 @diehlpk
@Muhao-Chen Could you please have a look here as well?
