### 2022-01-13 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @diehlpk
@Kevin-Mattheus-Moerman I think @Muhao-Chen worked on some of your remarks. Could you please have a look?
