+++ Software_Verification_and_Examples/Dynamics_Examples/03_double_pendulum/README.md
@@ -1,3 +1,7 @@
+This example is given as a unit test example. We comare the results betwwen the FEM method and rigid body approach. 
+
+The mass of the two bars in the double pendulum is m1 = m2 = m with a length of l1 = l2 = l. 
+
 This folder has two files:
 1. 'FEM_solution' is the dynamic analysis of the double pendulum by TsgFEM code.
 2. 'Analytcal_solution' is the analytical solution of the dynamic of the double pendulum.
