### 2022-01-21 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#1 @Muhao-Chen
@HaoZeke Thanks for your comments!
Please have a look at the  Tensegrity_Finite_Element_Method_TsgFEM/Software_Verification_and_Examples/Dynamics_Examples/03_double_pendulum/
Tensegrity_Finite_Element_Method_TsgFEM/Software_Verification_and_Examples/Dynamics_Examples/05_T_bar_Dynamics/
These simple examples are given as unit tests. For the double pendulum example, we also compare the results with rigid body dynamics. We also updated the comments in the codes and descriptions. Thank you!
