+++ Software_Verification_and_Examples/Dynamics_Examples/03_double_pendulum/README.md
@@ -1,5 +1,5 @@
 This example is given as a unit test example. We comare the results betwwen the FEM method and rigid body approach. 
-
+The initial configuration of the double pendulum is given as initial_config.png.
 The mass of the two bars in the double pendulum is m1 = m2 = m with a length of l1 = l2 = l. 
 
 This folder has two files:
