+++ JOSS_paper/paper.md
@@ -29,11 +29,15 @@ bibliography: paper.bib
 
 # Summary
 
-Tensegrity is a coined word by Buckminister Fuller [@fuller1982synergetics] for the art form created by Ioganson (1921) and Snelson (1948) [@lalvani1996origins] to represent a stable network of compressive members (bars/struts) and tensile members (strings/cables) [@chen2020general]. From the definition, it is straightforward to see that the fundamental property of the tensegrity is that all the bars and strings are axially loaded. Since the bars and strings are best in taking compression and tension and there is no material bending, the structure mass can be greatly reduced. Indeed, biological structures also indicate that tensegrity concepts yield the most efficient structures. For example, the bones, muscles, and elbows of animals and humans are tensegrity models. Wang et al. found that microtubules and microfilaments in the living cells work as compressive and tensile members to change the traction of the cell surfaces [@wang2001mechanical]. Simmons et al. showed that the DNA bundles are consistent with tensegrity prism [@liedl2010self]. After decades of study, many lightweight structures have been redesigned by the tensegrity paradigm. For example, Skelton and de Oliveira proved that T-Bar and D-Bar structures require less mass than a signal rod in taking compressive buckling load [@skelton2009tensegrity]. Ma et al. showed a mass efficient tensegrity cantilever structure subject to yield and buckling constraints [@ma2020design]. Barbarigos et al. designed and analyzed a lightweight pedestrian bridge [@rhode2010designing]. And many space applications are employing tensegrity solutions, for example, lightweight space habitat [@chen2021review], deployable lunar tower for space mining [@chen2020deployable], and planetary landers [@luo2017analysis;@sabelhaus2015system]. Moreover, the many advantages of tensegrity have also attracted researchers to find new ways to design soft robotics, i.e., six-bar tensegrity robot [@booth2020surface;@wang2020first], robotic spine [@sabelhaus2020model], morphing wings [@chen2020design], robotic fish [@chen2019swimming], debris capturing robot [@feng2021design].
+The name of this software, TsgFEM, is suggested to be pronounced as Tenseg FEM. TsgFEM provides two categories for the analysis of any tensegrity structure: statics and dynamics. For statics analysis, TsgFEM gives the minimum mass of the tensegrity structures by optimizing cross-sectional area in the strings and bars in the absence of external forces (self-equilibrium state) with any given external forces. The software allows solving for the minimum mass subject to buckling and yielding failures of structure members. The other important function of static analysis is that it can calculate the equilibrium configuration of tensegrity structures in any given external force, the forced motion of nodes, and the rest length change of strings. In the static analysis, the nonlinear properties of material and geometry can be considered, and the modified Newton method is used to guarantee the result converges to a stable equilibrium configuration. For the analysis of the dynamics, TsgFEM uses a second-order vector form differential equation to simulate the dynamics of any complexity of the tensegrity structure, including 1). Rigid body dynamics with acceptable errors. This is achieved by setting relatively high stiffness for bars in the dynamics simulation. 2). FEM dynamics simulation with elastic or plastic deformations in the presence of various kinds of boundary conditions, such as fix any nodes in any direction, apply static or dynamic external forces (i.e., gravitational force, some specified forces, or arbitrary seismic vibrations). 3). Accurate modal analysis, including natural frequency and corresponding modes. 4). An interface towards structural control by a compact state-space form of an accurate linear model.
 
[…]
