### 2022-01-21 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#1 @Muhao-Chen
@HaoZeke @diehlpk Thanks for pointing it out. We also updated the
Tensegrity_Finite_Element_Method_TsgFEM/User_Guide/User_Guide_Tensegrity_Finite_Element_Method_(TsgFEM) .pdf
As you may see, the unit tests examples, procedures, and results are described in detail in this pdf.  Thank you!
