### 2022-01-21 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#1 @diehlpk
@HaoZeke Please have a look at the new changes.
