+++ User_Guide/README.md
@@ -1,1 +1,2 @@
-We suggest new users read the attached pdf. Feedback and discussion are appreciated. 
+We suggest new users read the attached pdf carefully. This pdf contains all the steps in details in using this software. 
+Feedback and discussion are appreciated. Thank you! 
