+++ JOSS_paper/paper.md
@@ -29,9 +29,14 @@ bibliography: paper.bib
 
 # Summary
 
-The name of this software, TsgFEM, is suggested to be pronounced as Tenseg FEM. TsgFEM provides two categories for the analysis of any tensegrity structure: statics and dynamics. For statics analysis, TsgFEM gives the minimum mass of the tensegrity structures by optimizing cross-sectional area in the strings and bars in the absence of external forces (self-equilibrium state) with any given external forces. The software allows solving for the minimum mass subject to buckling and yielding failures of structure members. The other important function of static analysis is that it can calculate the equilibrium configuration of tensegrity structures in any given external force, the forced motion of nodes, and the rest length change of strings. In the static analysis, the nonlinear properties of material and geometry can be considered, and the modified Newton method is used to guarantee the result converges to a stable equilibrium configuration. For the analysis of the dynamics, TsgFEM uses a second-order vector form differential equation to simulate the dynamics of any complexity of the tensegrity structure, including 1). Rigid body dynamics with acceptable errors. This is achieved by setting relatively high stiffness for bars in the dynamics simulation. 2). FEM dynamics simulation with elastic or plastic deformations in the presence of various kinds of boundary conditions, such as fix any nodes in any direction, apply static or dynamic external forces (i.e., gravitational force, some specified forces, or arbitrary seismic vibrations). 3). Accurate modal analysis, including natural frequency and corresponding modes. 4). An interface towards structural control by a compact state-space form of an accurate linear model.
+The name of this software, TsgFEM, is suggested to be pronounced as Tenseg FEM. The purpose of this package is to facilitate the analysis of statics and dynamics of any tensegrity structures. The software allows one to perform topology design, statics and dynamic FEM simulation of any tensegrity systems. 
 
-# Motication and Introduction
+This software allows one to do the following studies, but not limited to the listed items. Topology design: 1). Modeling any tensegrity structures by nodal coordinates and the node's connectivity information. 2). Specifying the constraints of nodal coordinates, and grouping structure members and force them to have same force densities. Statics: 1). Conducting structure equilibrium configuration, prestress design, and stiffness studies. 2). Performing prestress and mechanism modes analysis. 3). Checking stiffness, stability, and robustness in terms of prestress, materials, and geometric information of the structure. 4). Conducting studies on form-finding of tensegrity systems. 5). Simulating the forced motion of structures. 6) Studying the feasibility of pseudo-static deployment trajectories. Dynamics: 1). Rigid body dynamics with acceptable errors. This is achieved by setting relatively high stiffness for bars in the dynamics simulation. 2). FEM dynamics simulation with elastic or plastic deformations in the presence of various kinds of boundary conditions, such as fix any nodes in any direction, apply static or dynamic external forces (i.e., gravitational force, some specified forces, or arbitrary seismic vibrations). 3). Modal analysis, including natural frequency and corresponding modes. 4). An interface towards structural control by a compact state-space form of a linear model.
+
+
[…]
