+++ JOSS_paper/paper.bib
@@ -167,13 +167,14 @@
   organization={IOP Publishing},
   doi={10.1088/1742-6596/1885/5/052024}
 }
-@misc{ma2021tensegrity,
- title={Tensegrity system dynamics based on finite element method}, 
- author={Shuo Ma and Muhao Chen and Robert E. Skelton},
- year={2021},
- eprint={2106.02176},
- archivePrefix={arXiv},
- primaryClass={physics.app-ph}
+@article{ma2021tensegrity,
+  title={Tensegrity system dynamics based on finite element method},
+  author={Ma, Shuo and Chen, Muhao and Skelton, Robert E},
+  journal={Composite Structures},
+  volume={280},
+  pages={114838},
+  year={2022},
+  publisher={Elsevier}
 }
 @inproceedings{zgoul2012static,
   title={Static Analysis of a Tensegrity Bridge using the Finite Element Method},
