### 2022-01-30 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#2 @Muhao-Chen
@HaoZeke @likask @diehlpk Thank you so much for your helpful comments! We have updated the paper.
https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/JOSS_paper/paper.md
Plus, we also have an updated user manual. All the procedures are in detailed steps. We believe this user manual would help users get comfortable with the software. Thank you!
https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/blob/main/User_Guide/User_Guide_Tensegrity_Finite_Element_Method_(TsgFEM)%20.pdf
