### 2022-02-14 #3390 @diehlpk
> @diehlpk Hello, I am happy with changes to the paper.

@likask

>
Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

Could you please have a look on this item? It seems not to be checked yet.
