### 2022-03-01 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
- [x] In `setup.m`, the following old syntax is still used producing a warning:
```matlab
clear all;
```
The new syntax is simply:
```matlab
clear;
```
- [x] In `ode4_tsg.m` and perhaps other codes/functions, code indentation is still not used, see comments above to fix this e.g.:
```matlab
if silentMode==0&useWaitbar==1
wb= waitbar(0,'Please wait...');
end
```
- [x] Also, check the above snippet e.g. if `silentMode==0&useWaitbar==1` should read `silentMode==0 && useWaitbar==1`
- [x] The file `Main_tower_static.m` features a capital `M` in `Main` while other codes have all lower case `main`. Consider making this consistent across all file names. Also check `Main_half_Tbar_static.m`, as well as all static demos e.g. `Main_lander.m`.
- [x] I get the following error in `main_Jasen_linkage_1.m` and similarly in `main_Jasen_linkage_2.m`, `main_Jasen_linkage_3.m`, and `main_cable_dome.m`:
```
Unrecognized function or variable 'tf'.

Error in main_Jasen_linkage_1 (line 151)
name=['Jasen_machanism','tf_',num2str(tf),material{1}];
```

- [x] Some codes (e.g. `main_Jasen_linkage_1.m`) still implement `figure(99);`. Implement the fix suggested above for all codes.
- [x] Various codes contain stuff like `ºÉÔØ×Ó²½` (e.g. `main_Jasen_linkage_1.m`). Are these perhaps language characters my computer does not know? Please review all codes and ensure all content works for an any international contributors.
```matlab
xlabel('ºÉÔØ×Ó²½','fontsize',14);
```
this produces the error:
```
Warning: Error updating Text.

 String scalar or character vector must have valid interpreter
 syntax:
ºÉÔØ×Ó²½
```

- [x] Please use `;` (e.g. `main_Jasen_linkage_1.m`) to terminate statements with outputs you do not which to have on the command screen currently. In `tenseg_plot_catenary` is used without a `;` causing the figure handle to be produced in the command window. Add semi-colons to suppress these. Check this in all codes.

- [x] The automated testing suggested has still not been added.
- [x] The code still feature non-meaningful or unspecified command window markers. Add words to say what these are where possible so the user knows what is been shown e.g.
```matlab
disp(['Current time: ',sprintf('%.5f',pi)]);
```

- [x] In `tenseg_ex_force.m` the input `a` is not used, one should remove it if not needed.
- [x] In `tenseg_ex_force.m` the switch statement says `'impluse'`, this should like be fixed to `'impulse'`
- [x] In `tenseg_ex_force.m` one could use `tspan<0.05` instead of `find(tspan<0.05)` to save time, i.e. there is not need to use `find`.
- [x]  In `tenseg_ex_force.m` there is not point to the empty array allocation `dz_a_t=[];`
- [x] In `tenseg_ex_force.m`, for the `'vib_force'` option, the variables `dz_d_t`, `dz_v_t`, and `dz_a_t`, are not used.
- [x] In `main_lander.m` check spelling of `dynaimcs`
- [x] In `main_lander.m`, none of the outputs of the following line (124) appear to be used:
 ```matlab
[w_t,dnb_t,dnb_d_t,dnb_dd_t,dz_a_t]=tenseg_ex_force(tspan,a,b,'impluse',gravity,[0;0;9.8],C,mass,3*(4:9)-1,0,period);
```

- [x] Question(s): The animation for the `main_lander.m` example appears to feature "contact" of the structure with a surface. Can you clarify how this is implemented? Is there documentation covering this type of boundary condition?

- [x] Also review the previously unchecked boxes above, and check occurrences of the points I make in other codes. Thanks.
