### 2022-03-01 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
@Muhao-Chen @diehlpk :wave: see above. I went through the changes. Some more changes needed.
