### 2022-03-01 #3390 @Kevin-Mattheus-Moerman
@diehlpk @Muhao-Chen Yes I responded over at https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/issues/3
