### 2022-03-02 85f361c
Update main_double_pendulum_simple.m
