+++ Function_library/ansys_input_gp.m
@@ -35,7 +35,7 @@ prestress_gp=t_gp./A_gp;
 fid11=fopen([name,'.txt'],'w');
 %%
 if language==0
-  % fprintf(fid11,'!输出节点坐标，拓扑关系，截面积，预应力\n\n');
+
 fprintf(fid11,'!nodal coordinate, topology, cross sectional area, prestress for ANSYS APDL\n\n');
 fprintf(fid11,'finish\n/clear \n/filename,tower  \n/title,the analysis of tower  \n!Unit:m，N，Pa，s\n\n');
 
@@ -46,34 +46,27 @@ fprintf(fid11,'!specify Youngs modulus \n es=%17.15f \n eg=%17.15f \n\n',Es,Eb);
 fprintf(fid11,'!specify string material property \n mp,ex,1,es	!Youngs modulus \n mp,prxy,1,0.3	!Poisson ratio\n mp,dens,1,%d	!Material density\nmp,alpx,1,6.5e-6	!coefficient of linear expansion\n\n',rho_s);
 fprintf(fid11,'!specify string material property \n mp,ex,2,eg	!Youngs modulus \n mp,prxy,2,0.3	!Poisson ratio\n mp,dens,2,%d	!Material density\nmp,alpx,2,6.5e-6	!coefficient of linear expansion\n\n',rho_b);
 
-% fprintf(fid11,'!定义单元类型 \n et,1,link180 \n \n');
-% fprintf(fid11,'!定义基本参数(数据) \n es=%17.15f \n eg=%17.15f \n\n',Es,Eb);
-% fprintf(fid11,'!定义材料特性(数据) \n mp,ex,1,es	!定义弹性模量 \n mp,prxy,1,0.3	!定义主泊松比\n mp,dens,1,%d	!定义质量密度\nmp,alpx,1,6.5e-6	!定义线膨胀系数\n\n',rho_s);
-% fprintf(fid11,'!定义材料特性(数据) \n mp,ex,2,eg	!定义弹性模量 \n mp,prxy,2,0.3	!定义主泊松比\n mp,dens,2,%d	!定义质量密度\nmp,alpx,2,6.5e-6	!定义线膨胀系数\n\n',rho_b);
-
 %% nodal coordinates,connectivity
 
 for i=1:nn
-%     fprintf(fid11,'K,%d,%17.15f,%17.15f,%17.15f  !节点坐标\n',i,N_xyz_nn(i,:));
+
     fprintf(fid11,'K,%d,%17.15f,%17.15f,%17.15f  !nodal coordinate\n',i,N_xyz_nn(i,:));
 end
 fprintf(fid11,'\n');
 for i=1:ne
-%     fprintf(fid11,'L,%4d,%4d  !创建线\n',El_123_nn(i,:));
+
     fprintf(fid11,'L,%4d,%4d  !line\n',El_123_nn(i,:));
 end
 fprintf(fid11,'\n');
 %% area
 fprintf(fid11,'*dim,area,,%d\n',ne);
 for i=1:ng
-%     fprintf(fid11,'area(%d)=%4d !截面积\n',i,A_gp(i));
+
     fprintf(fid11,'area(%d)=%4d !cross sectional area\n',i,A_gp(i));
 end
 fprintf(fid11,'\n');
-%  fprintf(fid11,'*DO,J,1,%d\n sectype,J,link  !定义截面类型为杆\nsecdata,area(J)	!定义杆截面几何数据\nseccontrol,,%d	!定义只拉不压（1）可拉可压（0）\n*ENDDO\n',ne,max(i==index_s));
 
 for i=1:ng
[…]
