### 2022-03-04 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#1 @HaoZeke
This has been resolved.
