### 2022-03-04 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#2 @HaoZeke
Thanks for addressing these points.
