### 2022-03-04 #3390 @diehlpk
> @diehlpk @Muhao-Chen Yes I responded over at [Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3](https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/issues/3)

@Muhao-Chen Please address the requested changes in a timely manner, so we can finalize the paper.
