+++ Software_Verification_and_Examples/Statics_Examples/04_Jasen_linkage/example1/main_Jasen_linkage_1.m
@@ -176,7 +176,7 @@ ylabel('λ��/m','fontsize',14)
 
 
 %% plot structure configuration
-tenseg_plot_catenary( reshape(n_t(:,end),3,[]),C_b,C_s,[],[],[0,0],[],[],l0_t(index_s,end))
+tenseg_plot_catenary( reshape(n_t(:,end),3,[]),C_b,C_s,[],[],[0,0],[],[],l0_t(index_s,end));
 % saveas(gcf,'1������̬.png');
 
 figure();
