+++ setup.m
@@ -5,20 +5,19 @@
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
 
-%% Add all necessary functions to MATLAB path%% set up the workspace
-clear;
-clc;
+%% Add all necessary functions to MATLAB path 
+% set up the workspace
+clear; clc;
 % add the function libraries
-addpath( genpath('Function_library') );
+addpath(genpath('Function_library'));
 % add the Software Verification and Examples
-addpath( genpath('Software_Verification_and_Examples') );
+addpath(genpath('Software_Verification_and_Examples'));
 % add User Guide
-addpath( genpath('User_Guide') );
+addpath(genpath('User_Guide'));
 % add Videos folder
-addpath( genpath('Videos') );
+addpath(genpath('Videos'));
 % add JOSS Paper
-addpath( genpath('JOSS_Paper') );
-
+addpath(genpath('JOSS_Paper'));
 %% Open the User_Guide
 % cd User_Guide;
 % open('User_Guide_Tensegrity_Finite_Element_Method_(TsgFEM).pdf');
