+++ README.md
@@ -4,46 +4,47 @@
 
 #### General Information
  
-Our group focuses the research of integrating structure and control design. Based on the tensegrity paradigm, we design tensegrity structures to meet the specified objectives. These objectives can vary from minimizing the mass of the structure to controlling the structure to meet certain performance. This software is intended to study the statics and dynamics of tensegrity systems based on FEM. The authors would like to make this as open-source software to help other researchers who are also interested in this field.
+Our group focuses on the research of integrating structure and control design. We design structures based on the tensegrity paradigm to meet the specified objectives. These objectives can vary from minimizing the structure's mass to controlling the structure to meet specific performance. This software is intended to study the statics and dynamics of tensegrity systems based on FEM. The authors would like to make this open-source software to help other researchers interested in this field.
 
-The purpose of this software is to facilitate the statics and dynamics of Tensegrity systems based on FEM.
-The software allows to perform modeling, structural design, nonlinear static and dynamic FEM simulation of any tensegrity and truss systems. Contribution of this software is mainly in these three aspects:
+This software aims to facilitate the statics and dynamics of Tensegrity systems based on FEM.
+The software allows modeling, structural design, nonlinear static, and dynamic FEM simulation of any tensegrity and truss systems. The contribution of this software is mainly in these three aspects.
+
 #### Modeling: 
-1. Model any tensegrity structures by nodal coordinates and the nodes’ connectivity information.
-2. Specify the constraints of nodal coordinates (restricted motions of the X-, X-, or/and Z-directions of some certain nodes). 
-3. Normally, structure members in symmetric positions have the same force densities, they are allowed to be grouped in the software. 
+1. Model any tensegrity structures by nodal coordinates and the nodes' connectivity information.
+2. Specify the constraints of nodal coordinates (restricted motions of the X-, X-, or/and Z-directions of some specific nodes). 
+3. Normally, structure members in symmetric positions have the same force densities. They are allowed to be grouped in the software. 
 #### Statics: 
 1. Prestress modes and mechanism modes analysis by singular value decomposition of equilibrium matrix.
 2. Prestress design and minimal mass calculation considering yielding and buckling constraints.
 3. Stiffness and stability analysis.
-4. Solve the equilibrium equations with any given external forces considering nonlinear of geometry and material.
-5. Simulate the forced motion by given sequencies of some nodes or changing rest length of some members.
+4. Solve the equilibrium equations with any given external forces considering the nonlinearity of geometry and material.
+5. Simulate the forced motion by giving sequences of some nodes or changing the rest length of some members.
 #### Dynamics: 
 1. Linearized dynamics.
 2. Modal analysis, calculate natural frequencies and mode shapes.
-3. Nonlinear dynamics simulation of any tensegrity structures or truss systems with linear elastic，multilinear elastic or plastic materials.
+3. Nonlinear dynamics simulation of any tensegrity structures or truss systems with linear elastic, multilinear elastic, or plastic materials.
 
 
 The name of this software _TsgFEM_ is suggested to be pronounced as _Tenseg FEM_. TsgFEM allows users to perform accurate dynamics and statics analysis for any tensegrity structures, including:
-1. Rigid body dynamics with acceptable errors. This is achieved by setting relatively high stiffness for bars in the dynamics simulation. 
-2. Actuate FEM dynamics simulation with elastic or plastic deformations in the presence of various kinds of boundary conditions, such as fix any nodes in any direction, apply static or dynamic external forces (i.e., gravitational force, some specified forces, or arbitrary seismic vibrations). 
[…]
