+++ Software_Verification_and_Examples/Statics_Examples/01_tower/tower_ansys.txt
@@ -30,11 +30,11 @@ K,1,10.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
 K,2,-4.999999999999998,8.660254037844387,0.000000000000000  !nodal coordinate
 K,3,-5.000000000000004,-8.660254037844386,0.000000000000000  !nodal coordinate
 K,4,8.660254037844386,5.000000000000000,30.000000000000000  !nodal coordinate
-K,5,-8.660254037844378,5.000000000000013,30.000000000000000  !nodal coordinate
-K,6,-0.000000000000004,-10.000000000000000,30.000000000000000  !nodal coordinate
+K,5,-8.660254037844386,5.000000000000004,30.000000000000000  !nodal coordinate
+K,6,-0.000000000000002,-10.000000000000000,30.000000000000000  !nodal coordinate
 K,7,4.999999999999999,8.660254037844387,60.000000000000000  !nodal coordinate
-K,8,-10.000000000000000,0.000000000000006,60.000000000000000  !nodal coordinate
-K,9,4.999999999999973,-8.660254037844402,60.000000000000000  !nodal coordinate
+K,8,-10.000000000000000,0.000000000000001,60.000000000000000  !nodal coordinate
+K,9,4.999999999999993,-8.660254037844391,60.000000000000000  !nodal coordinate
 
 L,   1,   5  !line
 L,   2,   6  !line

[…]
