### 2022-03-15 293a18c
Update Main_half_Tbar_static.m
