+++ Software_Verification_and_Examples/Dynamics_Examples/03_double_pendulum/Analytical_Solution/Main_Double_Pendulum_Analytical.m
@@ -3,7 +3,7 @@
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
 
-clc;clear all;close all;
+clc;clear;close all;
 % Initial Conditions
 m1 = 1; m2 = 1;
 L1 = 1; L2 = 1;
