+++ Software_Verification_and_Examples/Dynamics_Examples/05_T_bar_Dynamics/main_Tbar_dynamics.m
@@ -12,7 +12,7 @@
 % [3] dynamic simulation
 
 %EXAMPLE
-clc; clear ; close all;
+clc;clear;close all;
 % Global variable
 [consti_data,Eb,Es,sigmab,sigmas,rho_b,rho_s]=material_lib('Wood','Rubber_band');
 material{1}='linear_elastic'; % index for material properties: multielastic, plastic.
