+++ Function_library/equilibrium_solver.m
@@ -6,8 +6,8 @@ function data_out=equilibrium_solver(data,substep,slack,plastic)
 %converge to stable equilibrium, considering substep
 %%
 global E A l0 Ia Ib C w ne Xb Xa dXa
-% minimize total energy? (1: use, 0: not use) it's time consuming
-use_energy=0;
+% minimize total energy? (1: use, 0: not use). Using 1 is time consuming.
+use_energy = 0;
 
 switch nargin
     case 2
@@ -31,10 +31,10 @@ A=data.A;
 % w0=data.w;
 if  isfield(data,'w')
     if size(data.w,2)==substep
-                w_t=data.w;
+        w_t=data.w;
     elseif size(data.w,2)==1
-     w_t=data.w*linspace(0,1,substep);
-    end    
+        w_t=data.w*linspace(0,1,substep);
+    end
 else
     w_t=linspace(0,0,substep);
 end
@@ -42,19 +42,19 @@ end
 % dXb=data.dXb;
 if  isfield(data,'dXb')
     if size(data.dXb,2)==substep
-                dXb_t=data.dXb;
+        dXb_t=data.dXb;
     elseif size(data.dXb,2)==1
-     dXb_t=data.dXb*linspace(0,1,substep);
-    end    
+        dXb_t=data.dXb*linspace(0,1,substep);
+    end
 else
     dXb_t=linspace(0,0,substep);
 end
 % l0_0=data.l0;
-    if size(data.l0,2)==substep
-                l0_t=data.l0;
-    elseif size(data.l0,2)==1
-     l0_t=data.l0*linspace(1,1,substep);
-    end    
+if size(data.l0,2)==substep
+    l0_t=data.l0;
+elseif size(data.l0,2)==1
+    l0_t=data.l0*linspace(1,1,substep);
+end
 
 if  isfield(data,'subsubstep')
     subsubstep=data.subsubstep;
@@ -64,7 +64,7 @@ end
 
 X0=data.N(:);
 data_out=data;     %initialize output data
-    data_out.E_out=E0*ones(1,substep);
+data_out.E_out=E0*ones(1,substep);
 
 
 %% calculate equilibrium
@@ -72,7 +72,7 @@ X=X0;               %initialize configuration
 Xb0=Ib'*X;           %pinned node
 E=E0;
 % lamda=linspace(0,1,substep);    %coefficient for substep
- num_slack=ne*zeros(substep+1,1);    %num of string slack
+num_slack=ne*zeros(substep+1,1);    %num of string slack
 Xa=Ia'*X;
 cont=1;
 for k=1:substep
@@ -83,13 +83,13 @@ for k=1:substep
     u=1e-1;
     for i=1:1e3
         X=[Ia';Ib']\[Xa;Xb];
-        l=sqrt(sum((reshape(X,3,[])*C').^2))'; %bar length
-        q=E.*A.*(1./l0-1./l);      %force density
+        l=sqrt(sum((reshape(X,3,[])*C').^2))'; % bar length
+        q=E.*A.*(1./l0-1./l);      % force density
         q_bar=diag(q);
-        
-        K=kron(C'*q_bar*C,eye(3));                      %stiffness matrix
-        Fp=w-K*X;                                       %unbalanced force
-        Fp_a=Ia'*Fp;                                 %see the norm of unbalanced force
+
+        K=kron(C'*q_bar*C,eye(3));                      % stiffness matrix
+        Fp=w-K*X;                                       % unbalanced force
+        Fp_a=Ia'*Fp;                                    % see the norm of unbalanced force
         if norm(Fp_a)<1e-5
             break
         end
@@ -100,18 +100,18 @@ for k=1:substep
         end
[…]
