+++ Function_library/energy_cal.m
@@ -16,7 +16,6 @@ E=data.E;
 A=data.A;
 l0=data.l0;
 
-
-l=sqrt(sum((N*C').^2))'; %bar length 
-V=0.5*(l-l0)'*diag(E.*A./l0)*(l-l0)-w'*n;  %结构总势能
+l=sqrt(sum((N*C').^2))'; % bar length vector
+V=0.5*(l-l0)'*diag(E.*A./l0)*(l-l0)-w'*n;  % The total potential enegry of the structure
 end

[…]
