### 2022-03-15 18840ea
Update material_info.m
