+++ Function_library/material_info.m
@@ -8,8 +8,8 @@ function [consti_data,Eb,Es]=material_info()
 %% input the info of bars
 % strain_b1=[0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345 
 % stress_b1=1e6*[261,273,288.6,313.6,398.5,0,0];              % stress of bar Q345
-strain_b1=[1456e-6,23301e-6];  % strain of bar Q345  multi-linear
-stress_b1=1e6*[300,435];              % stress of bar Q345   multi-linear
+strain_b1=[1456e-6,23301e-6];  % strain of bar Q345, multi-linear
+stress_b1=1e6*[300,435];              % stress of bar Q345, multi-linear
 
 data_b0=[strain_b1;stress_b1];            %material info for strain>0
 data_b1=[-fliplr(data_b0),[0;0],data_b0];         %bar info from strain to stress
@@ -20,13 +20,13 @@ data_b_strain=strain_b(1:end-1);
 data_b2=[data_b_strain;data_b_E];               %bar info from strain to modulus
 
 %% input the info of strings
-% strain_s=[-1,0,0.016099,0.02,0.021];   % strain of string 钢丝绳
-% stress_s=1e6*[0,0,1223.5,0,0];          % stress of string 钢丝绳
-% stress_s=2*1e6*[0,0,790.95,0,0];          % stress of string 钢丝绳
+% strain_s=[-1,0,0.016099,0.02,0.021];   % strain of string, steel cable
+% stress_s=1e6*[0,0,1223.5,0,0];          % stress of string, steel cable
+% stress_s=2*1e6*[0,0,790.95,0,0];          % stress of string, steel cable
 % strain_s=[1,0,0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345 
 % stress_s=4*1e6*[0,0,261,273,288.6,313.6,398.5,0,0];              % stress of bar Q345
-strain_s=16*[-1,0,1456e-6,23301e-6];  % strain of bar Q345  双线性
-stress_s=4*1e6*[0,0,300,435];              % stress of bar Q345   双线性
+strain_s=16*[-1,0,1456e-6,23301e-6];  % strain of bar Q345, multi-linear
+stress_s=4*1e6*[0,0,300,435];              % stress of bar Q345, multi-linear
 
 data_s1=[strain_s;stress_s];                        %string info from strain to stress
 data_s_E=diff(stress_s)./diff(strain_s);
@@ -47,8 +47,8 @@ consti_data.data_s2=data_s2;
 figure
 plot(data_b1(1,:),data_b1(2,:),'k-o','linewidth',1.5)
 grid on; 
-xlabel('应变','fontsize',14);
-ylabel('应力/Pa','fontsize',14);title('杆');
+xlabel('Strain','fontsize',14);
+ylabel('Stress/Pa','fontsize',14);title('杆');
 % saveas(gcf,'1压杆本构.png');
 
 figure
