+++ Function_library/material_info.m
@@ -2,11 +2,11 @@ function [consti_data,Eb,Es]=material_info()
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
-% 
+%
 %output information of material: strain-stress, strain-modulus
 % bar is Q345 steel, string is steel rope
 %% input the info of bars
-% strain_b1=[0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345 
+% strain_b1=[0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345
 % stress_b1=1e6*[261,273,288.6,313.6,398.5,0,0];              % stress of bar Q345
 strain_b1=[1456e-6,23301e-6];  % strain of bar Q345, multi-linear
 stress_b1=1e6*[300,435];              % stress of bar Q345, multi-linear
@@ -23,7 +23,7 @@ data_b2=[data_b_strain;data_b_E];               %bar info from strain to modulus
 % strain_s=[-1,0,0.016099,0.02,0.021];   % strain of string, steel cable
 % stress_s=1e6*[0,0,1223.5,0,0];          % stress of string, steel cable
 % stress_s=2*1e6*[0,0,790.95,0,0];          % stress of string, steel cable
-% strain_s=[1,0,0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345 
+% strain_s=[1,0,0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345
 % stress_s=4*1e6*[0,0,261,273,288.6,313.6,398.5,0,0];              % stress of bar Q345
 strain_s=16*[-1,0,1456e-6,23301e-6];  % strain of bar Q345, multi-linear
 stress_s=4*1e6*[0,0,300,435];              % stress of bar Q345, multi-linear
@@ -45,43 +45,43 @@ consti_data.data_s2=data_s2;
 %% plot the stress-strain curve
 % stress_strain
 figure
-plot(data_b1(1,:),data_b1(2,:),'k-o','linewidth',1.5)
-grid on; 
+plot(data_b1(1,:),data_b1(2,:),'k-o','linewidth',1.5);
+grid on;
 xlabel('Strain','fontsize',14);
-ylabel('Stress/Pa','fontsize',14);title('杆');
-% saveas(gcf,'1压杆本构.png');
+ylabel('Stress/Pa','fontsize',14);
+% title('Bar');
+% saveas(gcf,'bar_constitutive.png');
 
 figure
-plot(data_s1(1,:),data_s1(2,:),'k-o','linewidth',1.5)
+plot(data_s1(1,:),data_s1(2,:),'k-o','linewidth',1.5);
 grid on;
-axis([-0.015,0.025,-max(stress_s),max(stress_s)])
-xlabel('应变','fontsize',14);
-ylabel('应力/Pa','fontsize',14);
-% title('索');
-% saveas(gcf,'1拉索本构.png');
+axis([-0.015,0.025,-max(stress_s),max(stress_s)]);
+xlabel('Strain','fontsize',14);
+ylabel('Stress/Pa','fontsize',14);
+% title('String');
+% saveas(gcf,'string_constitutive.png');
 %% plot the modulus-strain curve
 % E-strain
 figure
 xi = linspace(2*min(data_b_strain),2*max(data_b_strain),1000);
[…]
