### 2022-03-15 baaf003
Update material_info2.m
