### 2022-03-15 557d71e
Update material_info2.m
