+++ Function_library/material_info2.m
@@ -7,12 +7,12 @@ function [consti_data,Eb,Es]=material_info2()
 % bar is Q345 steel, string is steel rope.
 
 %% input the info of bars
-% strain_b1=[0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345 
+% strain_b1=[0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345
 % stress_b1=1e6*[261,273,288.6,313.6,398.5,0,0];              % stress of bar Q345
-strain_b1=[1456e-6,23301e-6,30301e-6];  % strain of bar Q345 
-stress_b1=1e6*[300,435,435];              % stress of bar Q345  
+strain_b1=[1456e-6,23301e-6,30301e-6];  % strain of bar Q345
+stress_b1=1e6*[300,435,435];              % stress of bar Q345
 
-data_b0=[strain_b1;stress_b1];            % material info for strain>0
+data_b0=[strain_b1;stress_b1];            % material info for strain > 0
 data_b1=[-fliplr(data_b0),[0;0],data_b0];         % bar info from strain to stress
 stress_b=data_b1(2,:);
 strain_b=data_b1(1,:);
@@ -24,8 +24,8 @@ data_b2=[data_b_strain;data_b_E];               %bar info from strain to modulus
 strain_s=[-1,0,0.016099,1];   % strain of string, steel cable
 stress_s=1e6*[0,0,1223.5,1223.5];          % stress of string, steel cable
 % stress_s=2*1e6*[0,0,790.95,0,0];          % stress of string, steel cable
-% strain_s=[1,0,0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345 
-% stress_s=4*1e6*[0,0,261,273,288.6,313.6,398.5,0,0];              % stress of bar Q345
+% strain_s=[1,0,0.001267,0.00157,0.02480,0.033,0.06,0.07,0.08];  % strain of bar Q345
+% stress_s=4*1e6*[0,0,261,273,288.6,313.6,398.5,0,0];            % stress of bar Q345
 
 
 data_s1=[strain_s;stress_s];                        %string info from strain to stress
@@ -46,9 +46,9 @@ consti_data.data_s2=data_s2;
 % stress_strain
 figure
 plot(data_b1(1,:),data_b1(2,:),'k-o','linewidth',1.5)
-grid on; 
+grid on;
 axis([-0.03,0.03,-1.2*max(stress_b),1.2*max(stress_b)]);
-xlabel('应变','fontsize',14);
+xlabel('Strain','fontsize',14);
 ylabel('应力/Pa','fontsize',14);
 title('杆');
 % saveas(gcf,'1压杆本构.png');
@@ -57,7 +57,7 @@ figure
 plot(data_s1(1,:),data_s1(2,:),'k-o','linewidth',1.5)
 grid on;
 axis([-0.015,0.025,-1.2*max(stress_s),1.2*max(stress_s)]);
-xlabel('应变','fontsize',14);
+xlabel('Strain','fontsize',14);
 ylabel('应力/Pa','fontsize',14);
 title('索');
 % saveas(gcf,'1拉索本构.png');
@@ -68,23 +68,21 @@ xi = linspace(2*min(data_b_strain),2*max(data_b_strain),100000);
 yi = interp1(data_b_strain, data_b_E, xi,'previous',0);
 plot(data_b_strain, data_b_E,'ko', xi, yi,'linewidth',1.5);
[…]
