### 2022-03-15 9199e35
Update static_solver.m
