+++ Function_library/static_solver.m
@@ -6,7 +6,7 @@ function data_out=static_solver(data)
 % solve nonlinear equilibrium equations using modified Newton method
 % converge to stable equilibrium, considering substep
 %%
-global E A l0 Ia Ib C w ne Xb Xa dXa f_int l_int 
+global E A l0 Ia Ib C w ne Xb Xa dXa f_int l_int
 % minimize total energy? (1: use, 0: not use) it's time consuming
 use_energy=1;
 
@@ -25,10 +25,10 @@ A=data.A;
 % w0=data.w;
 if  isfield(data,'w_t')
     if size(data.w_t,2)==substep
-                w_t=data.w_t;
+        w_t=data.w_t;
     elseif size(data.w_t,2)==1
-     w_t=data.w_t*linspace(0,1,substep);
-    end    
+        w_t=data.w_t*linspace(0,1,substep);
+    end
 else
     w_t=linspace(0,0,substep);
 end
@@ -36,65 +36,65 @@ end
 % dXb=data.dXb;
 if  isfield(data,'dnb_t')
     if size(data.dnb_t,2)==substep
-                dXb_t=data.dnb_t;
+        dXb_t=data.dnb_t;
     elseif size(data.dnb_t,2)==1
-     dXb_t=data.dnb_t*linspace(0,1,substep);
-    end    
+        dXb_t=data.dnb_t*linspace(0,1,substep);
+    end
 else
     dXb_t=linspace(0,0,substep);
 end
 % l0_0=data.l0;
-    if size(data.l0_t,2)==substep
-                l0_t=data.l0_t;
-    elseif size(data.l0_t,2)==1
-     l0_t=data.l0_t*linspace(1,1,substep);
-    end    
+if size(data.l0_t,2)==substep
+    l0_t=data.l0_t;
+elseif size(data.l0_t,2)==1
+    l0_t=data.l0_t*linspace(1,1,substep);
+end
 
 if  isfield(data,'subsubstep')
     subsubstep=data.subsubstep;
 else
-    subsubstep=30;          %default ssubstep
+    subsubstep=30;          % default ssubstep
 end
 
 X0=data.N(:);
-data_out=data;     %initialize output data
-    data_out.E_out=E0*ones(1,substep);
+data_out=data;     % initialize output data
+data_out.E_out=E0*ones(1,substep);
 
 
 %% calculate equilibrium
-X=X0;               %initialize configuration
-Xb0=Ib'*X;           %pinned node
+X=X0;               % initialize configuration
+Xb0=Ib'*X;           % pinned node
 E=E0;
-% lamda=linspace(0,1,substep);    %coefficient for substep
- num_slack=ne*zeros(substep+1,1);    %num of string slack
+% lamda=linspace(0,1,substep);    % coefficient for substep
+num_slack=ne*zeros(substep+1,1);    % number of string slack
 Xa=Ia'*X;
 cont=2;
 for k=1:substep
-    w=w_t(:,k);               %external force
-    Xb=Xb0+dXb_t(:,k);         %forced node displacement
-    l0=l0_t(:,k);         %forced enlongation of string
+    w=w_t(:,k);               % external force
+    Xb=Xb0+dXb_t(:,k);         % forced node displacement
+    l0=l0_t(:,k);         % forced enlongation of string
     disp(k);
     u=1e-1;
-    
-%     X=[Ia';Ib']\[Xa;Xb];
-%     l=sqrt(sum((reshape(X,3,[])*C').^2))'; %bar length
-%     strain=(l-l0)./l0;        %strain of member
-%     [E,sigma]=stress_strain(consti_data,index_b,index_s,strain,material);
[…]
