### 2022-03-15 111294c
Update static_solver2.m
