+++ Function_library/static_solver2.m
@@ -110,11 +110,11 @@ for k=1:substep
         K_taa=Ia'*K_t*Ia;
         
         %modify the stiffness matrix
-        [V_mode,D]=eig(K_taa);                       %刚度矩阵特征根
-        d=diag(D);                            %eigen value
-        lmd=min(d);                     %刚度矩阵最小特征根
+        [V_mode,D]=eig(K_taa);                       % eigenvalues of the stiffness matrix
+        d=diag(D);                            %eigenvalues
+        lmd=min(d);                     % the smallest eigenvalue
         if lmd>0
-            Km=K_taa+u*eye(size(K_taa)); %修正的刚度矩阵
+            Km=K_taa+u*eye(size(K_taa)); % modified stiffness matrix
         else
             Km=K_taa+(abs(lmd)+u)*eye(size(K_taa));
         end
@@ -182,7 +182,6 @@ for k=1:substep
     
     
     %% output data
-    
     data_out.N_out{k}=reshape(X,3,[]);
     data_out.n_out(:,k)=X;
     %     data_out.l_out(:,k)=l;
