+++ Function_library/tenseg_prestress_design.m
@@ -1,12 +1,12 @@
 function [q_gp,t_gp,q,t]=tenseg_prestress_design(Gp,l,l_gp,A_1ag,V2,w0a,index_gp,fd)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
-% * file, You can obtain one at http://mozilla.org/MPL/2.0/. 
+% * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 %
 % This function gives the prestress design of tensegrity, consider boundary
 % constraints and group constraints, given deisgned force( vector with
 % number of prestress modes entries)
-% 
+%
 % Inputs:
 %   Gp: group matrix of members
 %   l: members' length vector
@@ -20,9 +20,9 @@ function [q_gp,t_gp,q,t]=tenseg_prestress_design(Gp,l,l_gp,A_1ag,V2,w0a,index_gp
 % Outputs:
 %	q_gp: force desity vector in group
 %	t_gp: force vector in group
-%	q: force desity vector 
-%	t: force vector 
-%% 
+%	q: force desity vector
+%	t: force vector
+%%
 I=eye(size(Gp,2));
 e_d=I(:,index_gp);        % e_d is the matrix to select group of member with designed force
 l_d=e_d'*l_gp;            % length of top center circular strings
@@ -38,4 +38,3 @@ q=q1+q2;                        % force density
 t=diag(l)*q;                    % force vector
 t_gp=pinv(Gp)*t;                % force in group
 end
-
