+++ Software_Verification_and_Examples/Statics_Examples/05_cable_dome/generat_cable_dome.m
@@ -15,7 +15,6 @@ function [N,C_b,C_s,C] =generat_cable_dome(R,p,m,h,beta)
 %	N: nodal coordinate matrix
 %   C_b,C_s: connectivity matrix of bars and strings
 %   C: connectivity matrix of all members
-
 %% nodal coordinate
 l=2*R;
 Rd=(h^2+(l/2)^2)/(2*h); %radius of the arch
@@ -35,11 +34,10 @@ T1=[cos(beta1) -sin(beta1) 0
     0 0 1];
 for i=1:m    %rotate nodes
     if rem(i,2)==1
-    N0(:,i)=T1* N0(:,i);
-     N0(:,i+m)=T1* N0(:,i+m);
+        N0(:,i)=T1* N0(:,i);
+        N0(:,i+m)=T1* N0(:,i+m);
     end
 end
-
 %% replicate units
 beta2=2*pi/p;
 T2=[cos(beta2) -sin(beta2) 0
@@ -49,10 +47,8 @@ N=N0;
 for i=1:p-1
     N=[N,T2^i*N0];
 end
-
-
 %% connectivity matrix of a unit
-C_b_in_unit = [[1:m]',[m+1:2*m]'];  % Bar 
+C_b_in_unit = [[1:m]',[m+1:2*m]'];  % Bar
 % Convert above index notation into actual connectivity matrices
 C_b2unit = tenseg_ind2C(C_b_in_unit,N);
 
@@ -74,37 +70,37 @@ C_s_in_temp=[1 2  % strings of inner loop
     m+3 m+3+u];%2 top 2 diagonal 1 circlur 2 top 2 diagonal 1 circlur
 
 if rem(m,2)==1     %odd m
-C_s_in_unit=[C_s_in_unit1;[kron(2*[0:(m-1)/2-1]',ones(10,2))+kron(ones((m-1)/2,1),C_s_in_temp)]];
-C_s_in_unit=[C_s_in_unit;m,m+u];     % top circular strings
+    C_s_in_unit=[C_s_in_unit1;[kron(2*[0:(m-1)/2-1]',ones(10,2))+kron(ones((m-1)/2,1),C_s_in_temp)]];
+    C_s_in_unit=[C_s_in_unit;m,m+u];     % top circular strings
 else if m==2
-    C_s_in_unit=[1 u   % 2 top 2 diagonal 1 circlur
-    1 2*u
-    m+1 u
-    m+1 2*u
-    m+1 m+1+u
-    1 2  
-    1 2+u
-    1 m+2
-    1 m+2+u
-    m+2 m+2+u
-    m,m+u];
+        C_s_in_unit=[1 u   % 2 top 2 diagonal 1 circlur
+            1 2*u
+            m+1 u
+            m+1 2*u
+            m+1 m+1+u
+            1 2
+            1 2+u
+            1 m+2
+            1 m+2+u
+            m+2 m+2+u
+            m,m+u];
 else if rem(m,2)==0&&m~=2
-C_s_in_unit=[C_s_in_unit1;[kron(2*[0:(m-2)/2-2]',ones(10,2)),kron(ones((m-2)/2,1),C_s_in_temp)]];
-C_s_in_unit=[C_s_in_unit;           
-    m-1 m  
-    m-1 m+u
-    m-1 2*m
-    m-1 2*m+u
-    2*m 2*m+u
-m,m+u];     % top circular strings
+        C_s_in_unit=[C_s_in_unit1;[kron(2*[0:(m-2)/2-2]',ones(10,2)),kron(ones((m-2)/2,1),C_s_in_temp)]];
+        C_s_in_unit=[C_s_in_unit;
+            m-1 m
+            m-1 m+u
+            m-1 2*m
[…]
