### 2022-03-15 e06c00a
Update Main_half_Tbar_static.m
