### 2022-03-15 6bb67a7
Update tenseg_plot_catenary.m
