### 2022-03-15 ca8a825
Update energy_cal.m
