### 2022-03-15 d881be3
Update equilibrium_solver.m
