### 2022-03-15 16c36a7
Update equilibrium_solver.m
