+++ Function_library/equilibrium_solver.m
@@ -2,8 +2,8 @@ function data_out=equilibrium_solver(data,substep,slack,plastic)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
-%solve nonlinear equilibrium equations using modified Newton method
-%converge to stable equilibrium, considering substep
+% solve nonlinear equilibrium equations using modified Newton method
+% converge to stable equilibrium, considering substep
 %%
 global E A l0 Ia Ib C w ne Xb Xa dXa
 % minimize total energy? (1: use, 0: not use). Using 1 is time consuming.
@@ -13,10 +13,10 @@ switch nargin
     case 1
         substep = 1;
         slack = 0;
-        plastic=0;
+        plastic = 0;
     case 2
         slack = 0;
-        plastic=0;
+        plastic = 0;
 end
 %% input data
 C=data.C;
@@ -58,30 +58,30 @@ end
 if  isfield(data,'subsubstep')
     subsubstep=data.subsubstep;
 else
-    subsubstep=30;          %default ssubstep
+    subsubstep=30;          % default substep
 end
 
 X0=data.N(:);
-data_out=data;     %initialize output data
+data_out=data;     % initialize output data
 data_out.E_out=E0*ones(1,substep);
 %% calculate equilibrium
-X=X0;               %initialize configuration
-Xb0=Ib'*X;           %pinned node
+X=X0;                % initialize configuration
+Xb0=Ib'*X;           % pinned nodes
 E=E0;
-% lamda=linspace(0,1,substep);    %coefficient for substep
-num_slack=ne*zeros(substep+1,1);    %num of string slack
+% lamda=linspace(0,1,substep);      % coefficient for the substeps
+num_slack=ne*zeros(substep+1,1);    % number of the slack strings
 Xa=Ia'*X;
 cont=1;
 for k=1:substep
-    w=w_t(:,k);               %external force
-    Xb=Xb0+dXb_t(:,k);         %forced node displacement
-    l0=l0_t(:,k);         %forced enlongation of string
+    w=w_t(:,k);                % external force
+    Xb=Xb0+dXb_t(:,k);         % forced displacement of nodes
+    l0=l0_t(:,k);              % forced enlongation of the strings
     disp(k);
     u=1e-1;
     for i=1:1e3
         X=[Ia';Ib']\[Xa;Xb];
         l=sqrt(sum((reshape(X,3,[])*C').^2))'; % bar length
-        q=E.*A.*(1./l0-1./l);      % force density
+        q=E.*A.*(1./l0-1./l);                  % force density
         q_bar=diag(q);
 
         K=kron(C'*q_bar*C,eye(3));                      % stiffness matrix
@@ -98,9 +98,9 @@ for k=1:substep
         K_t=kron(C',eye(3))*blkdiag(Ki{:})*kron(C,eye(3));
         K_taa=Ia'*K_t*Ia;
 
-        %modify the stiffness matrix
-        [V_mode,D]=eig(K_taa);                       % eigenvalues of the stiffness matrix
-        d=diag(D);                            % eigenvalues
+        % modify the stiffness matrix
+        [V_mode,D]=eig(K_taa);          % eigenvalues of the stiffness matrix
+        d=diag(D);                      % eigenvalues
         lmd=min(d);                     % the smallest eigenvalue
         if lmd>0
             Km=K_taa+u*eye(size(K_taa)); % modified stiffness matrix
@@ -110,7 +110,7 @@ for k=1:substep
         dXa=Km\Fp_a;
 
         x=1;
-        % line search
[…]
