+++ Function_library/material_lib.m
@@ -30,7 +30,7 @@ global  Eb Es
 switch bar_material
     case 'Steel_Q345'
         %         rho_b = 7870;
-        rho_b = 7870*50;    %amplified density
+        rho_b = 7870*50;    % amplified density
         strain_b1=[1456e-6,23301e-6,1];  % strain of bar Q345
         stress_b1=1e6*[300,435,435];              % stress of bar Q345
         [data_b1,data_b2,Eb,sigma_b]=point2consti_data(strain_b1,stress_b1);
@@ -102,7 +102,7 @@ switch string_material
         %---------------- Aluminum for Strings -------------------
         Es = 60e09;
         %         rho_s = 2700;
-        rho_s = 2700*50;    %amplified density not real
+        rho_s = 2700*50;    % amplified density not real
         sigma_s = 110e06;
         [data_s1,data_s2]=blin_consti_data(Es,sigma_s);
     case 'Rubber_band'
