### 2022-03-15 c8fae6c
Delete plot_data_torus.m
