+++ Function_library/plot_data_torus.m
@@ -1,712 +1,0 @@
-function plot_data_torus(data_out,direction,saveimg)
-% /* This Source Code Form is subject to the terms of the Mozilla Public
-% * License, v. 2.0. If a copy of the MPL was not distributed with this
-% * file, You can obtain one at http://mozilla.org/MPL/2.0/.
-%
-%plot the information of torus R=10
-%%
-N_out=data_out.N_out;
-t_step=data_out.t_out;          %member force in every step
-n_step=data_out.n_out;          %nodal coordinate in every step
-substep=size(n_step,2);
-%% plot member force and node coordinate
-
-% plot_data_torus(data_out,direction,saveimg);
-zb=1:substep;
-if direction==1
-    figure
-    plot(zb,t_step(147,:),'k-o',zb,t_step(152,:),'k-^','linewidth',1.5);
-    set(gca,'fontsize',18);
-    legend('外脊索','内脊索')
-    xlabel('荷载子步','fontsize',18);
-    ylabel('内力/N','fontsize',18);
-    if saveimg==1
-        saveas(gcf,'1脊索内力.png');
-    end
-
-    figure
-    plot(zb,t_step(149,:),'k-o',zb,t_step(154,:),'k-^','linewidth',1.5);
-    set(gca,'fontsize',18);
-    legend('外斜索','内斜索','location','northwest')
-    xlabel('荷载子步','fontsize',18);
-    ylabel('内力/N','fontsize',18);
-    if saveimg==1
-        saveas(gcf,'1斜索内力.png');
-    end
-
-    figure
-    plot(zb,t_step(372,:),'k-o',zb,t_step(156,:),'k-^',zb,t_step(157,:),'k-v','linewidth',1.5);
-    set(gca,'fontsize',18);
-    legend('外环索','内环索','内环顶索','location','northwest')
-    xlabel('荷载子步','fontsize',18);
-    ylabel('内力/N','fontsize',18);
-    if saveimg==1
-        saveas(gcf,'1环索内力.png');
-    end
-
-    figure
-    plot(zb,t_step(145,:),'k-o',zb,t_step(146,:),'k-^','linewidth',1.5);
-    set(gca,'fontsize',18);
-    legend('外竖杆','内竖杆')
-    xlabel('荷载子步','fontsize',18);
-    ylabel('内力/N','fontsize',18);
-    if saveimg==1
-        saveas(gcf,'1竖杆内力.png');
-    end
-
-    figure
-    plot(zb,t_step(4,:),'k-o',zb,t_step(19,:),'k-^','linewidth',1.5);
-    set(gca,'fontsize',18);
-    legend('水平杆','稳定杆')
-    xlabel('荷载子步','fontsize',18);
-    ylabel('内力/N','fontsize',18);
-    if saveimg==1
-        saveas(gcf,'1环杆内力.png');
-    end
-
-    figure
-    plot(zb,t_step(55,:),'k-o',zb,t_step(76,:),'k-^','linewidth',1.5);
-    set(gca,'fontsize',18);
-    legend('环竖索','环斜索')
-    xlabel('荷载子步','fontsize',18);
-    ylabel('内力/N','fontsize',18);
[…]
