+++ Function_library/ansys_input_mode.m
@@ -1,4 +1,4 @@
-function[]= ansys_input(N,C,A,t,b,index_s,name)
+function[]= ansys_input_mode(N,C,A,t,b,index_s,name)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -10,129 +10,118 @@ for i=1:ne
     El_123_nn(i,:)=find(C(i,:));
 end
 prestress=t./A;
-
-
 %% output for ansys
 fid11=fopen(name,'w');
-%% 
-  fprintf(fid11,'!输出节点坐标，拓扑关系，截面积，预应力\n!材料特性，弹塑性需要自己更改\n\n');
- fprintf(fid11,'finish\n/clear \n/filename,tower  \n/title,the analysis of tower  \n!单位m，N，Pa，s\n\n');
- 
- %%  material properties
-  fprintf(fid11,'/prep7\n');
- fprintf(fid11,'!定义单元类型 \n et,1,link180 \n \n');
- fprintf(fid11,'!定义基本参数(数据) \n es=76000e6 \n eg=2.06e11 \nfd=2.15e08 \nfy=2.35e08 \n ft=1.57e09 \n ratio1=0.85 \nratio2=0.70\n');
- fprintf(fid11,'!定义材料特性(数据) \n mp,ex,1,es	!定义弹性模量 \n mp,prxy,1,0.3	!定义主泊松比\n mp,dens,1,7870	!定义质量密度\nmp,alpx,1,6.5e-6	!定义线膨胀系数\n\n');
- fprintf(fid11,'!定义材料特性(数据) \n mp,ex,2,eg	!定义弹性模量 \n mp,prxy,2,0.3	!定义主泊松比\n mp,dens,2,7870	!定义质量密度\nmp,alpx,2,6.5e-6	!定义线膨胀系数\n\n');
+%%
+fprintf(fid11,'!输出节点坐标，拓扑关系，截面积，预应力\n!材料特性，弹塑性需要自己更改\n\n');
+fprintf(fid11,'finish\n/clear \n/filename,tower  \n/title,the analysis of tower  \n!单位m，N，Pa，s\n\n');
+
+%%  material properties
+fprintf(fid11,'/prep7\n');
+fprintf(fid11,'!定义单元类型 \n et,1,link180 \n \n');
+fprintf(fid11,'!定义基本参数(数据) \n es=76000e6 \n eg=2.06e11 \nfd=2.15e08 \nfy=2.35e08 \n ft=1.57e09 \n ratio1=0.85 \nratio2=0.70\n');
+fprintf(fid11,'!定义材料特性(数据) \n mp,ex,1,es	!定义弹性模量 \n mp,prxy,1,0.3	!定义主泊松比\n mp,dens,1,7870	!定义质量密度\nmp,alpx,1,6.5e-6	!定义线膨胀系数\n\n');
+fprintf(fid11,'!定义材料特性(数据) \n mp,ex,2,eg	!定义弹性模量 \n mp,prxy,2,0.3	!定义主泊松比\n mp,dens,2,7870	!定义质量密度\nmp,alpx,2,6.5e-6	!定义线膨胀系数\n\n');
 
 %% nodal coordinates,connectivity
 
- for i=1:nn
-     fprintf(fid11,'K,%d,%17.15f,%17.15f,%17.15f  !节点坐标\n',i,N_xyz_nn(i,:));
- end
-  fprintf(fid11,'\n');
[…]
