### 2022-03-15 1eeb198
Update generat_N_kuihua.m
