### 2022-03-15 577c278
Update main_tower_dynamic_simple.m
