+++ Software_Verification_and_Examples/Dynamics_Examples/01_tower/tjx.txt
@@ -1,2001 +1,2001 @@
-0.0000000000000000000000000  
-0.6283019941676303288247141  
-1.2565047721668740621225879  
-1.8845091334967269958156066  
-2.5122159089884776861367754  
-3.1395259764656686662931406  
-3.7663402763966362662984011  
-4.3925598275371582346338073  
-5.0180857425607445776449822  
-5.6428192436740847170995039  
-6.2666616782152129161431731  
-6.8895145342319024095445457  
-7.5112794560378528530009135  
-8.1318582597441810833061027  
-8.7511529487638011204353461  
-9.3690657292862304927894002  
-9.9854990257203510850558814  
-10.6003554961027326442035701  
-11.2135380474690595775655311  
-11.8249498511862345395684315  
-12.4344943582427394090927919  
-13.0420753144948466939467835  
-13.6475967758662584117246297  
-14.2509631234988063397395308  
-14.8520790788517462743811848  
-15.4508497187473690104297930  
-16.0471804903604784442450182  
-16.6409772261493316136693466  
-17.2321461587258539793765522  
-17.8205939356625364666797395  
-18.4062276342338968504463992  
-18.9889547760900541106821038  
-19.5686833418601224821031792  
-20.1453217856831301446618454  
-20.7187790496642065818377887  
-21.2889645782536369722492964  
-21.8557883325466484336629946  
-22.4191608045016117500836117  
-22.9789930310743919505966915  
-23.5351966082666272939150076  
-24.0876837050857659505709307  
-24.6363670774145795405729586  
-25.1811600817880396618875238  
-25.7219766890753192001284333  
-26.2587314980647796858193033  
-26.7913397489498343873037811  
-27.3197173367134560351132677  
-27.8437808244093965015508729  
-28.3634474563378269351687777  
-28.8786351711133768560557655  
-29.3892626146236572992620495  
-29.8952491528759409789017809  
-30.3965148847302728540853423  
-30.8929806545167160436449194  
-31.3845680645350242343738500  
-31.8711994874344810568800312  
-32.3527980784722117846285983  
-32.8292877876478286225392367  
-33.3005933717125799375935458  
-33.7666404060512235218993737  
-34.2273552964344318638723053  
-34.6826652906402443932165625  
-35.1324984899424563877801120  
-35.5767838604642676614275842  
-36.0154512443953436218180286  
-36.4484313710705762900943228  
-36.8756558679086978713712597  
-37.2970572712091055223027070  
-37.7125690368051849077346560  
-38.1221255505723917167415493  
-38.5256621387894639951809950  
-38.9231150783511736790387658  
-39.3144216068309475531350472  
-39.6995199323917660194638302  
-40.0783492435438262191382819  
-40.4508497187473707867866324  
-40.8169625358591972030808392  
-41.1766298814213698165076494  
-41.5297949597906281837822462  
-41.8764020021070848542876774  
-42.2163962751007559859317553  
-42.5497240897345889720781997  
-42.8763328096826157320720085  
-43.1961708596417608418960299  
-43.5091877334762813234192436  
-43.8153340021931825276624295  
-44.1145613217476650902426627  
-44.4068224406772316115166177  
-44.6920712075631811899256718  
-44.9702625783185467867042462  
-45.2413526233009761767789314  
-45.5052985342497819942764181  
-45.7620586310458747902885079  
-46.0115923682935132887905638  
-46.2538603417229055025927664  
-46.4888242944125664735111059  
-46.7164471228306013017572695  
-46.9366928826937055418966338  
-47.1495267946432221606301027  
-47.3549152497372176640055841  
-47.5528258147576750047846872  
-47.7432272373321495706477435  
-47.9260894508687940174240794  
-48.1013835793042971999966539  
[…]
