+++ Software_Verification_and_Examples/Dynamics_Examples/05_T_bar_Dynamics/main_Tbar_dynamics.m
@@ -137,7 +137,7 @@ data.M=M;data.D=D;
 data.tf=tf;data.dt=dt;data.tspan=tspan;data.out_tspan=out_tspan;
 
 %% plot external force information
-tenseg_plot_exforce(Ib,tspan,w_t,[1],dnb_t,dnb_d_t,dnb_dd_t,[1],saveimg);
+% tenseg_plot_exforce(Ib,tspan,w_t,[1],dnb_t,dnb_d_t,dnb_dd_t,[1],saveimg);
 
 %% dynamic analysis 
 % solve dynamic equation
