+++ Software_Verification_and_Examples/Statics_Examples/02_3D_Dbar/modeling_3d_Dbar/three_d_bar.m
@@ -7,11 +7,10 @@ function [N,C_b,C_s] = three_d_bar(p1,p2,q,angle)
 [N,C_b_n,C_s_n]=threedbar(p1,p2,angle,q);
 % save all_test
 % [y,N,C_b_in,C_s_in]=convert_index_connectivity([1 0 0 0],pi/6,2);
-% 
+%
 % N = round(N,5);
 % C_b_n = round(C_b_n,5);
 % C_s_n = round(C_s_n,5);
-
 N = setoff_dup(N,6);
 C_b_in = transfer_C_b(N,C_b_n);
 C_s_in = transfer_C_b(N,C_s_n);
