+++ Software_Verification_and_Examples/Statics_Examples/02_3D_Dbar/modeling_3d_Dbar/setoff_dup.m
@@ -8,12 +8,13 @@ function N = setoff_dup(N,precision)
 %       precision: Integer for precision
 %  Outputs:
 %       N: N*3 matrix with no duplicates
+%%
 index = [];
 for i=1:length(N(:,1))
     a = find(((N(:,1)-N(i,1)).^2 + (N(:,2)-N(i,2)).^2 +(N(:,3)-N(i,3)).^2).^0.5<10^-precision ==1)
     tmpa = a(find(a>i));
     if ~isempty(tmpa)
-    index = [index;tmpa];
+        index = [index;tmpa];
     end
 end
 N(index,:)=[];
