+++ Software_Verification_and_Examples/Statics_Examples/04_Jasen_linkage/example1/main_Jasen_linkage_1.m
@@ -5,7 +5,6 @@
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 %%
-clc;clear;close all;
 % global l  Eb Es
 %EXAMPLE
 clc; clear ; close all;
