+++ Software_Verification_and_Examples/Statics_Examples/04_Jasen_linkage/example1/main_Jasen_linkage_1.m
@@ -7,7 +7,7 @@
 %%
 % global l  Eb Es
 %EXAMPLE
-clc; clear ; close all;
+clc; clear; close all;
 % Global variable
 [consti_data,Eb,Es,sigmab,sigmas,rho_b,rho_s]=material_lib('Wood','Steel_string');
 material{1}='linear_elastic'; % index for material properties:'linear_elastic' multielastic plastic
@@ -24,7 +24,7 @@ saveimg=0;              % save image or not (1) yes (0)no
 savedata=0;             % save data or not (1) yes (0)no
 savevideo=1;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
-
+tf = 1/substep;         % simulation time step
 %% N C of the structure
 % Manually specify node positions (accurate).
 N=[   38.0000   38.0000   -8.7357  -39.6678         0  -19.4476   17.0047   30.3109
@@ -125,7 +125,7 @@ n_t=data_out.n_out;          % nodal coordinate in every step
 N_out=data_out.N_out;
 
 %% Plot final configuration
-% tenseg_plot_catenary(N_out{end},C_b,C_s,[],[],[0,90],[],[],l0_t(index_s,end))
+% tenseg_plot_catenary(N_out{end},C_b,C_s,[],[],[0,90],[],[],l0_t(index_s,end));
 
 %% input file of ANSYS
 ansys_input_gp(N_out{1},C,A_gp,t_gp,b,Eb,Es,rho_b,rho_s,Gp,index_s,find(t_gp>0),'Jasen_mechanism1');
@@ -137,7 +137,7 @@ tenseg_plot_result(1:substep,t_t(1:3,:),{'element 1','element 2','element 3'},{'
 % tenseg_plot_result(1:substep,n_t([3*8-2,3*8-1],:),{'8X','8Y'},{'Time (s)','Coordinate (m)'},'X_coordinate.png',saveimg);
 
 %% Plot final configuration
-% tenseg_plot_catenary( reshape(n_t(:,end),3,[]),C_b,C_s,[],[],[0,0],[],R3Ddata,l0_t(index_s,end))
+% tenseg_plot_catenary(reshape(n_t(:,end),3,[]),C_b,C_s,[],[],[0,0],[],R3Ddata,l0_t(index_s,end));
 
 %% save output data
 if savedata==1
@@ -154,5 +154,3 @@ tenseg_video_slack(n_t,C_b,C_s,l0_t,index_s,[],[0,90],[-80,80,-85,50,0,70],min(s
 
 %% plot structure configuration
 % tenseg_plot_catenary( reshape(n_t(:,end),3,[]),C_b,C_s,[],[],[0,90],[],[],l0_t(index_s,end));
-
-
