+++ Software_Verification_and_Examples/Dynamics_Examples/truss_10seg2.txt.txt
@@ -1,441 +1,0 @@
-!nodal coordinate, topology, cross sectional area, prestress for ANSYS APDL
-
-finish
-/clear 
-/filename,tower  
-/title,the analysis of tower  
-!Unit:m，N，Pa，s
-
-/prep7
-!specify element type 
- et,1,link180 
- 
-!specify Youngs modulus 
- es=75998509224.175415039062500 
- eg=206043956043.956024169921875 
-
-!specify string material property 
- mp,ex,1,es	!Youngs modulus 
- mp,prxy,1,0.3	!Poisson ratio
- mp,dens,1,7870	!Material density
-mp,alpx,1,6.5e-6	!coefficient of linear expansion
-
-!specify string material property 
- mp,ex,2,eg	!Youngs modulus 
- mp,prxy,2,0.3	!Poisson ratio
- mp,dens,2,7870	!Material density
-mp,alpx,2,6.5e-6	!coefficient of linear expansion
-
-K,1,0.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,2,1.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,3,2.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,4,3.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,5,4.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,6,5.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,7,6.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,8,7.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,9,8.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,10,9.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,11,10.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,12,0.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,13,1.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,14,2.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,15,3.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,16,4.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,17,5.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,18,6.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,19,7.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,20,8.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,21,9.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-K,22,10.000000000000000,1.000000000000000,0.000000000000000  !nodal coordinate
-
-L,   1,   2  !line
-L,   2,   3  !line
-L,   3,   4  !line
-L,   4,   5  !line
-L,   5,   6  !line
-L,   6,   7  !line
-L,   7,   8  !line
-L,   8,   9  !line
-L,   9,  10  !line
-L,  10,  11  !line
-L,   1,  12  !line
-L,   1,  13  !line
-L,   2,  13  !line
-L,   2,  14  !line
-L,   3,  14  !line
-L,   3,  15  !line
-L,   4,  15  !line
-L,   4,  16  !line
[…]
