+++ Software_Verification_and_Examples/Statics_Examples/3d_Dbar_ansys.txt
@@ -1,153 +1,0 @@
-!nodal coordinate, topology, cross sectional area, prestress for ANSYS APDL
-
-finish
-/clear 
-/filename,tower  
-/title,the analysis of tower  
-!Unit:m，N，Pa，s
-
-/prep7
-!specify element type 
- et,1,link180 
- 
-!specify Youngs modulus 
- es=75998509224.175415039062500 
- eg=206043956043.956024169921875 
-
-!specify string material property 
- mp,ex,1,es	!Youngs modulus 
- mp,prxy,1,0.3	!Poisson ratio
- mp,dens,1,7870	!Material density
-mp,alpx,1,6.5e-6	!coefficient of linear expansion
-
-!specify string material property 
- mp,ex,2,eg	!Youngs modulus 
- mp,prxy,2,0.3	!Poisson ratio
- mp,dens,2,393500	!Material density
-mp,alpx,2,6.5e-6	!coefficient of linear expansion
-
-K,1,0.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,2,6.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
-K,3,3.000000000000000,0.000000000000000,1.091910702798607  !nodal coordinate
-K,4,3.000000000000000,-0.945622407287714,-0.545955351399304  !nodal coordinate
-K,5,3.000000000000000,0.945622407287714,-0.545955351399304  !nodal coordinate
-
-L,   1,   3  !line
-L,   1,   5  !line
-L,   1,   4  !line
-L,   2,   3  !line
-L,   2,   5  !line
-L,   2,   4  !line
-L,   3,   4  !line
-L,   4,   5  !line
-L,   3,   5  !line
-L,   1,   2  !line
-
-*dim,area,,10
-area(1)=1.494710e-03 !cross sectional area
-area(2)=1.145012e-05 !cross sectional area
-area(3)=8.173273e-05 !cross sectional area
-
-sectype,1,link  !specify section type
-secdata,area(1)   !specify section data
-seccontrol,,0       !only in tension(1) both tension and compression(0) 
-sectype,2,link  !specify section type
-secdata,area(2)   !specify section data
-seccontrol,,1       !only in tension(1) both tension and compression(0) 
-sectype,3,link  !specify section type
-secdata,area(3)   !specify section data
-seccontrol,,1       !only in tension(1) both tension and compression(0) 
-
-!define element type
-lsel,s,,,1  !select element
-latt,2,,1,,,,1  !specify section area
-lsel,s,,,2  !select element
-latt,2,,1,,,,1  !specify section area
-lsel,s,,,3  !select element
-latt,2,,1,,,,1  !specify section area
-lsel,s,,,4  !select element
-latt,2,,1,,,,1  !specify section area
-lsel,s,,,5  !select element
-latt,2,,1,,,,1  !specify section area
-lsel,s,,,6  !select element
-latt,2,,1,,,,1  !specify section area
-lsel,s,,,7  !select element
-latt,1,,1,,,,2  !specify section area
-lsel,s,,,8  !select element
-latt,1,,1,,,,2  !specify section area
-lsel,s,,,9  !select element
-latt,1,,1,,,,2  !specify section area
-lsel,s,,,10  !select element
-latt,1,,1,,,,3  !specify section area
-
-*dim,prestress,,10
[…]
