+++ Software_Verification_and_Examples/Dynamics_Examples/03_double_pendulum/FEM_solution/tower.txt
@@ -1,0 +1,179 @@
+!nodal coordinate, topology, cross sectional area, prestress for ANSYS APDL
+
+finish
+/clear 
+/filename,tower  
+/title,the analysis of tower  
+!Unit:m，N，Pa，s
+
+/prep7
+!specify element type 
+ et,1,link180 
+ 
+!specify Youngs modulus 
+ es=75998509224.175415039062500 
+ eg=206043956043.956024169921875 
+
+!specify string material property 
+ mp,ex,1,es	!Youngs modulus 
+ mp,prxy,1,0.3	!Poisson ratio
+ mp,dens,1,7870	!Material density
+mp,alpx,1,6.5e-6	!coefficient of linear expansion
+
+!specify string material property 
+ mp,ex,2,eg	!Youngs modulus 
+ mp,prxy,2,0.3	!Poisson ratio
+ mp,dens,2,393500	!Material density
+mp,alpx,2,6.5e-6	!coefficient of linear expansion
+
+K,1,10.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
+K,2,-4.999999999999998,8.660254037844387,0.000000000000000  !nodal coordinate
+K,3,-5.000000000000004,-8.660254037844386,0.000000000000000  !nodal coordinate
+K,4,8.660254037844386,5.000000000000000,30.000000000000000  !nodal coordinate
+K,5,-8.660254037844386,5.000000000000004,30.000000000000000  !nodal coordinate
+K,6,-0.000000000000002,-10.000000000000000,30.000000000000000  !nodal coordinate
+K,7,4.999999999999999,8.660254037844387,60.000000000000000  !nodal coordinate
+K,8,-10.000000000000000,0.000000000000001,60.000000000000000  !nodal coordinate
+K,9,4.999999999999993,-8.660254037844391,60.000000000000000  !nodal coordinate
+
+L,   1,   5  !line
+L,   2,   6  !line
+L,   3,   4  !line
+L,   5,   9  !line
+L,   6,   7  !line
+L,   4,   8  !line
+L,   4,   5  !line
+L,   5,   6  !line
+L,   4,   6  !line
+L,   7,   8  !line
+L,   8,   9  !line
+L,   7,   9  !line
+L,   1,   4  !line
+L,   2,   5  !line
+L,   3,   6  !line
+L,   4,   7  !line
+L,   5,   8  !line
+L,   6,   9  !line
+
+*dim,area,,18
+area(1)=8.870014e-02 !cross sectional area
+area(2)=8.870014e-02 !cross sectional area
+area(3)=4.581175e-04 !cross sectional area
+area(4)=2.290587e-04 !cross sectional area
+area(5)=6.973305e-04 !cross sectional area
+area(6)=6.973305e-04 !cross sectional area
+
+sectype,1,link  !specify section type
+secdata,area(1)   !specify section data
+seccontrol,,0       !only in tension(1) both tension and compression(0) 
+sectype,2,link  !specify section type
+secdata,area(2)   !specify section data
+seccontrol,,0       !only in tension(1) both tension and compression(0) 
+sectype,3,link  !specify section type
+secdata,area(3)   !specify section data
+seccontrol,,1       !only in tension(1) both tension and compression(0) 
+sectype,4,link  !specify section type
+secdata,area(4)   !specify section data
+seccontrol,,1       !only in tension(1) both tension and compression(0) 
+sectype,5,link  !specify section type
+secdata,area(5)   !specify section data
[…]
