+++ Function_library/equilibrium_solver_pinv.m
@@ -1,4 +1,4 @@
-function data_out=nonlinequ(data)
+function data_out=equilibrium_solver_pinv(data)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
