+++ Function_library/generat_N_kuihua.m
@@ -22,7 +22,7 @@ beta1=pi/p;
 T1=[cos(beta1) -sin(beta1) 0
     sin(beta1) cos(beta1) 0
     0 0 1];
-for i=1:m    %rotate nodes
+for i=1:m    % rotate nodes
     if rem(i,2)==1
         N0(:,i)=T1* N0(:,i);
         N0(:,i+m)=T1* N0(:,i+m);
