### 2022-03-17 7156bbc
test statics
