+++ Software_Verification_and_Examples/Statics_Examples/04_Jasen_linkage/example3/Jasen linkage full size.txt
@@ -29,51 +29,51 @@ mp,alpx,2,6.5e-6	!coefficient of linear expansion
 K,1,0.000000000000000,7.800000000000000,0.000000000000000  !nodal coordinate
 K,2,0.000000000000000,22.800000000000001,0.000000000000000  !nodal coordinate
 K,3,-46.735652302443299,40.570166118107245,0.000000000000000  !nodal coordinate
-K,4,-77.667791263174934,-5.871655328881544,0.000000000000000  !nodal coordinate
+K,4,-77.667791263174934,-5.871655328881545,0.000000000000000  !nodal coordinate
 K,5,-38.000000000000000,0.000000000000000,0.000000000000000  !nodal coordinate
 K,6,-57.447599367531701,-39.687388940668882,0.000000000000000  !nodal coordinate
 K,7,-20.995300642707409,-35.430639279698198,0.000000000000000  !nodal coordinate
-K,8,-7.689066230641667,-82.589351367404305,0.000000000000000  !nodal coordinate
+K,8,-7.689066230641668,-82.589351367404305,0.000000000000000  !nodal coordinate
 K,9,0.000000000000000,7.800000000000000,-10.000000000000000  !nodal coordinate
 K,10,-12.990381056766580,0.300000000000001,-10.000000000000000  !nodal coordinate
-K,11,20.351349892921998,37.560286865222842,-10.000000000000000  !nodal coordinate
-K,12,73.106342950595078,19.379233334553859,-10.000000000000000  !nodal coordinate
+K,11,20.351349892921991,37.560286865222842,-10.000000000000000  !nodal coordinate
+K,12,73.106342950595092,19.379233334553863,-10.000000000000000  !nodal coordinate
 K,13,38.000000000000000,0.000000000000000,-10.000000000000000  !nodal coordinate
-K,14,65.555693978951368,-19.290492551558994,-10.000000000000000  !nodal coordinate
+K,14,65.555693978951368,-19.290492551558984,-10.000000000000000  !nodal coordinate
 K,15,34.700498194921309,-39.161247271227396,-10.000000000000000  !nodal coordinate
-K,16,54.384410280714015,-84.033773442580454,-10.000000000000000  !nodal coordinate
+K,16,54.384410280714043,-84.033773442580440,-10.000000000000000  !nodal coordinate
 K,17,0.000000000000000,7.800000000000000,-20.000000000000000  !nodal coordinate
 K,18,12.990381056766575,0.299999999999994,-20.000000000000000  !nodal coordinate
-K,19,-20.351349892922006,37.560286865222849,-20.000000000000000  !nodal coordinate
+K,19,-20.351349892922002,37.560286865222849,-20.000000000000000  !nodal coordinate
 K,20,-73.106342950595092,19.379233334553856,-20.000000000000000  !nodal coordinate
 K,21,-38.000000000000000,0.000000000000000,-20.000000000000000  !nodal coordinate
-K,22,-65.555693978951368,-19.290492551558994,-20.000000000000000  !nodal coordinate
+K,22,-65.555693978951382,-19.290492551558998,-20.000000000000000  !nodal coordinate
 K,23,-34.700498194921316,-39.161247271227396,-20.000000000000000  !nodal coordinate
-K,24,-54.384410280714036,-84.033773442580454,-20.000000000000000  !nodal coordinate
+K,24,-54.384410280714015,-84.033773442580468,-20.000000000000000  !nodal coordinate
 K,25,0.000000000000000,7.800000000000000,-30.000000000000000  !nodal coordinate
 K,26,0.000000000000000,22.800000000000001,-30.000000000000000  !nodal coordinate
[…]
