+++ Software_Verification_and_Examples/Statics_Examples/01_tower/tower_ansys.txt
@@ -4,7 +4,7 @@ finish
 /clear 
 /filename,tower  
 /title,the analysis of tower  
-!Unit:m，N，Pa，s
+!Unit:m£¬N£¬Pa£¬s
 
 /prep7
 !specify element type 
@@ -146,7 +146,7 @@ latt,1,,1,,,,6  !specify section area
 LMESH,ALL
 finish
 
-!First solve for self-equilibrium）
+!First solve for self-equilibrium£©
 /SOLU
 ANTYPE,0 
 NLGEO!consider large deformation 
