### 2022-03-17 4ad0b51
Update test_Statics.m
