### 2022-03-17 28f0b1f
Update test_Statics.m
