+++ Software_Verification_and_Examples/README.md
@@ -1,1 +1,3 @@
-This folder contains dynamics and static examples of TsgFEM. To better understand this software, we suggest user first read these two files: 'tower' and 'lander' in the Statics_Examples and Dynamics_Examples folders.
+This folder contains dynamics and static examples of TsgFEM. 
+To better understand this software, we suggest user first read these two files: 'tower' and 'lander' in the Statics_Examples and Dynamics_Examples folders.
+To simplify the process, the user can also run the auto_statics_test.m and auto_dynamics_test.m to perform simulation examples.
