### 2022-03-17 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Muhao-Chen
@Kevin-Mattheus-Moerman @diehlpk Thank you so much for all the valuable suggestions. We followed your advice and made changes. Please have a check and let me know if you have any comments. Thank you!!

I would also like to respond to some of the pointed issues.
> * [x]  Various codes contain stuff like `ºÉÔØ×Ó²½` (e.g. `main_Jasen_linkage_1.m`). Are these perhaps language characters my computer does not know? Please review all codes and ensure all content works for any international contributors.
Thank you for pointing it out! Yes, we have international collaborators. So we provided both English and Chinese descriptions. Because the language setting of your Windows operating system is not in Chinese, so you see the wield codes. To solve this issue, we provided both language options for users. As you may see, in the help description of function ansys_input_gp.m, there is a language option: language=0;   % 0 for English,1 for Chinese. For elsewhere, we have changed all the descriptions into English.
>
> * [x]  The automated testing suggested has still not been added.
Yes, we have followed your provided procedure and added this function. Very cool and thank you!
> * [x]  Question(s): The animation for the `main_lander.m` example appears to feature "contact" of the structure with a surface. Can you clarify how this is implemented? Is there documentation covering this type of boundary condition?
> The ground is modeled as a second-order system with ground stiffness and damping. The function and description are given at here: Tensegrity_Finite_Element_Method_TsgFEM/Software_Verification_and_Examples/Dynamics_Examples/04_Lander/w_t_ground_force.m.
