### 2022-03-28 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @diehlpk
@Kevin-Mattheus-Moerman Could you please have a look?
