### 2022-04-05 22a689c
Fixed filesep symbol to be linux compatible, added more tests, cleanup static test script
