+++ Software_Verification_and_Examples/auto_statics_test.m
@@ -15,15 +15,24 @@
 % Below is to run the statics examples automatically.
 %%
 clear; close all; clc;
-%%
-testSet='all';
-testMode='test';
-approveQuestion=0;
-startLoc=1;
+
 %% running tests
 originFolder=fileparts(mfilename('fullpath'));
-testFileList={'Main_tower_static.m','Main_folding_3d_Dbar.m','Main_lander.m','main_Jasen_linkage_3.m'};
-testFolderList={'01_tower','02_3D_Dbar','03_lander','04_Jasen_linkage\example3'}
+testFileList={'Main_tower_static.m',...
+              'Main_folding_3d_Dbar.m',...
+              'Main_lander.m',...
+              'main_Jasen_linkage_1.m',...
+              'main_Jasen_linkage_2.m',...
+              'main_Jasen_linkage_3.m',...
+              };
+testFolderList={'01_tower',...
+                '02_3D_Dbar',...
+                '03_lander',...
+                ['04_Jasen_linkage',filesep,'example1'],...
+                ['04_Jasen_linkage',filesep,'example2'],...
+                ['04_Jasen_linkage',filesep,'example3'],...
+                }; 
+
 %% loop over all examples
 for q_test=1:1:numel(testFileList)
 
