+++ Software_Verification_and_Examples/auto_statics_test.m
@@ -1,8 +1,9 @@
 %% Automated test 
-% The auomated test file format is from *GIBBON footer text*. 
-% The source code is here:
+
+% This automated test code was derived from codes from the GIBBON project.
+% The source code is here: 
 % https://github.com/gibbonCode/GIBBON/blob/master/lib/testGibbon.m
-% The license of _*GIBBON footer text*_ is given below. 
+% The license for GIBBON is given below. 
 % License: <https://github.com/gibbonCode/GIBBON/blob/master/LICENSE>
 % And the copyright information is:
 % Copyright (C) 2006-2021 Kevin Mattheus Moerman and the GIBBON
@@ -11,6 +12,7 @@
 % along with this program.  If not, see <http://www.gnu.org/licenses/>.
 % Please include their licence and Copyright information when you use this
 % file. 
+
 %% test Statics
 % Below is to run the statics examples automatically.
 %%
@@ -21,16 +23,16 @@ originFolder=fileparts(mfilename('fullpath'));
 testFileList={'Main_tower_static.m',...
               'Main_folding_3d_Dbar.m',...
               'Main_lander.m',...
-              'main_Jasen_linkage_1.m',...
-              'main_Jasen_linkage_2.m',...
-              'main_Jasen_linkage_3.m',...
+              'main_Jansen_linkage_1.m',...
+              'main_Jansen_linkage_2.m',...
+              'main_Jansen_linkage_3.m'...
               };
 testFolderList={'01_tower',...
                 '02_3D_Dbar',...
                 '03_lander',...
-                ['04_Jasen_linkage',filesep,'example1'],...
-                ['04_Jasen_linkage',filesep,'example2'],...
-                ['04_Jasen_linkage',filesep,'example3'],...
+                ['04_Jansen_linkage',filesep,'example1'],...
+                ['04_Jansen_linkage',filesep,'example2'],...
+                ['04_Jansen_linkage',filesep,'example3']...
                 }; 
 
 %% loop over all examples
