+++ Software_Verification_and_Examples/Statics_Examples/04_Jansen_linkage/example3/main_Jansen_linkage_3.m
@@ -1,0 +1,181 @@
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%%%%%%%%%%%%%%%Jansen linkage with 3 parts%%%%%%%%%%%%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+% /* This Source Code Form is subject to the terms of the Mozilla Public
+% * License, v. 2.0. If a copy of the MPL was not distributed with this
+% * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+%% EXAMPLE
+clc; clear; close all;
+% Global variable
+[consti_data,Eb,Es,sigmab,sigmas,rho_b,rho_s]=material_lib('Wood','Steel_string');
+material{1}='linear_elastic'; % index for material properties:'linear_elastic' multielastic plastic
+material{2}=0; % index for considering slack of string (1) for yes,(0) for no (for compare with ANSYS)
+
+% cross section design cofficient
+thick=6e-3;        % thickness of hollow bar
+hollow_solid=0;    % use hollow bar or solid bar in minimal mass design (1)hollow (0)solid
+c_b=0.1;           % coefficient of safty of bars 0.5
+c_s=0.1;           % coefficient of safty of strings 0.3
+
+substep=100;                                     %substep
+lumped=0;               % use lumped matrix 1-yes,0-no
+saveimg=0;              % save image or not (1) yes (0)no
+savedata=0;             % save data or not (1) yes (0)no
+savevideo=1;            % make video(1) or not(0)
+gravity=0;              % consider gravity 1 for yes, 0 for no
+tf = 1/substep;         % simulation time step
+%% N C of the structure
+% Manually specify node positions (accurate).
+d_z=10;
+N0=[   38.0000   38.0000   -8.7357  -39.6678         0  -19.4476   17.0047   30.3109
+    7.8000   22.8000   40.5702   -5.8717         0  -39.6874  -35.4306  -82.5894
+    0         0         0         0         0         0         0         0];
+N0(1,:)=N0(1,:)-38;
+N1=N0;N1(3,:)=-d_z*ones(1,8);N1(1,:)=-N1(1,:);
+N2=N0;N2(3,:)=-2*d_z*ones(1,8);
+N3=N0;N3(3,:)=-3*d_z*ones(1,8);N3(1,:)=-N3(1,:);
+N4=N0;N4(3,:)=-4*d_z*ones(1,8);
+N5=N0;N5(3,:)=-5*d_z*ones(1,8);N5(1,:)=-N5(1,:);
+N=[N0,N1,N2,N3,N4,N5];
+% N=kron([1 1 1],N0);
+
+% Manually specify connectivity indices.
+C_s_in=[];  % This is indicating that string connection
+C_b_in0 = [1 2;2 3;3 4;3 5;4 5;4 6;5 7;6 7;6 8;7 8;2 7];  % Similarly, this is saying bar 1 connects node 1 to node 2,
+C_b_in=[C_b_in0; C_b_in0+8; C_b_in0+16; C_b_in0+24; C_b_in0+32; C_b_in0+40];
+% Convert the above matrices into full connectivity matrices.
[…]
