### 2022-04-05 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
Most things are now fixed, please review and consider merging this pull request to fix remaining issues:
https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/pull/5
