### 2022-04-06 bec3df2
Fixed minor bugs, made sure codes do not produce output in working directory
