+++ Software_Verification_and_Examples/Statics_Examples/04_Jansen_linkage/example3/main_Jansen_linkage_3.m
@@ -24,6 +24,8 @@ savedata=0;             % save data or not (1) yes (0)no
 savevideo=1;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
 tf = 1/substep;         % simulation time step
+savePath=fileparts(mfilename('fullpath')); %Save files in same folder as this code
+
 %% N C of the structure
 % Manually specify node positions (accurate).
 d_z=10;
@@ -137,10 +139,10 @@ n_t=data_out.n_out;          %nodal coordinate in every step
 N_out=data_out.N_out;
 
 %% input file of ANSYS
-ansys_input_gp(N_out{1},C,A_gp,t_gp,b,Eb,Es,rho_b,rho_s,Gp,index_s,find(t_gp>0),'Jansen linkage full size');
+ansys_input_gp(N_out{1},C,A_gp,t_gp,b,Eb,Es,rho_b,rho_s,Gp,index_s,find(t_gp>0),fullfile(savePath,'Jansen linkage full size'));
 
 %% plot member force
-tenseg_plot_result(1:substep,t_t(1:3,:),{'element 1','element 2','element 3'},{'Load step','Force (N)'},'member_force.png',saveimg);
+tenseg_plot_result(1:substep,t_t(1:3,:),{'element 1','element 2','element 3'},{'Load step','Force (N)'},fullfile(savePath,'member_force.png'),saveimg);
 
 %% Plot nodal coordinate curve X Y
 % tenseg_plot_result(1:substep,n_t([3*8-2,3*8-1],:),{'8X','8Y'},{'Time (s)','Coordinate (m)'},'X_coordinate.png',saveimg);
@@ -148,8 +150,13 @@ tenseg_plot_result(1:substep,t_t(1:3,:),{'element 1','element 2','element 3'},{'
 %% Plot final configuration
 % tenseg_plot_catenary( reshape(n_t(:,end),3,[]),C_b,C_s,[],[],[0,90],[],R3Ddata,l0_t(index_s,end));
 
+%% save output data
+if savedata==1
+    save (fullfile(savePath,['jansen_linkage',material{1},'.mat']));
+end
+
 %% make video of the dynamic
-name=['Jansen_machanism3','tf_',num2str(tf),material{1}];
+name=fullfile(savePath,['Jansen_machanism3','tf_',num2str(tf),material{1}]);
 % tenseg_video(n_t,C_b,C_s,[],substep,name,savevideo);
 tenseg_video_slack(n_t,C_b,C_s,l0_t,index_s,[],[0,90],[],min(substep,50),name,savevideo,material{2});
 
