+++ Software_Verification_and_Examples/Statics_Examples/03_lander/main_lander.m
@@ -1,5 +1,5 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-%%%%%%%%%%%%%%%Static analysis of folding a 3d Dbar %%%%%%%%%%%%%%%%%%%%%%%%
+%%%%%%%%%%%%%%%Static calculation of a tensegrity tower %%%%%%%%%%%%%%%%%%%%
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -21,8 +21,9 @@
 clc;clear;close all;
 % Specify material properties
 [consti_data,Eb,Es,sigmab,sigmas,rho_b,rho_s]=material_lib('Steel_Q345','Steel_string');
-material{1}='linear_elastic'; % index for material properties:linear_elastic; multielastic; plastic
+material{1}='linear_elastic'; % index for material properties:'linear_elastic' multielastic plastic
 material{2}=0; % index for considering slack of string (1) for yes,(0) for no (for compare with ANSYS)
+
 % cross section design cofficient
 thick=6e-3;        % thickness of hollow bar
 hollow_solid=0;          % use hollow bar or solid bar in minimal mass design (1)hollow (0)solid
@@ -35,25 +36,38 @@ saveimg=0;              % save image or not (1) yes (0)no
 savedata=0;             % save data or not (1) yes (0)no
 savevideo=1;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
+savePath=fileparts(mfilename('fullpath')); %Save files in same folder as this code
 
 %% N C of the structure
-% Manually specify node positions of a 3D D-Bar Unit.
-angle = pi/180*20;
-sizes=4;
-q=1;
-[N,C_b,C_s]=three_d_bar([0,0,0],[6,0,0],q,angle);
+% Manually specify node positions of a lander.
+L = 1;
+th1 = -atan(1/2);
+N = [L/4 0 0;L/4 0 L;-L/4 0 0;-L/4 0 L;
+    0 -L/2 3*L/4;0 L/2 3*L/4;0 -L/2 L/4;0 L /2 L/4;
+    L/2 -L/4 L/2;-L/2 -L/4 L/2;L/2 L/4 L/2;-L/2 L/4 L/2]';
+N = [1 0 0;0 cos(th1) -sin(th1);0 sin(th1) cos(th1)]*N;
+N(3,:)=N(3,:)+2;
+% N = [cos(th2) -sin(th2) 0;sin(th2) cos(th2) 0;0 0 1]*N;
+C_b_in = [1 2;3 4;5 6;7 8;9 10;11 12];
+C_s_in = [2 5;2 6;2 9;2 11;4 5;4 6;4 10;4 12;
+    1 7;1 8;1 9;1 11;3 7;3 8;3 10;3 12;
+    5 9;5 10;7 9;7 10;6 11;6 12;8 11;8 12];
+% Convert the above matrices into full connectivity matrices.
+C_b = tenseg_ind2C(C_b_in,N);%%
+C_s = tenseg_ind2C(C_s_in,N);
 C=[C_b;C_s];
 [ne,nn]=size(C);        % ne:No.of element;nn:No.of node
 % Plot the structure to make sure it looks right
 tenseg_plot(N,C_b,C_s);
 
 %% Boundary constraints
[…]
