### 2022-04-06 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
@Muhao-Chen
- [x] I created one more pull request for you to consider: https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/pull/6
- [x] The function ansys_input_gp_00 still contains what looks like non-english symbols my system cannot parse. Please update.
See line 45 `m£¬N£¬Pa£¬s`, line 149 `!Êä³ö½Úµã×ø±ê£¬ÍØÆË¹ØÏµ£¬½ØÃæ»ý£¬Ô¤Ó¦Á¦`, and lines 155-158 e.g. `¨Òå²ÄÁÏÌØÐÔ(Êý¾Ý)`, line 163 `!½Úµã×ø±ê` ....and many others.
- [x] functions like `tenseg_mode` and `tenseg_video_strain_stress` save images to the current working directory. Change this throughout to be a local folder. Also using git to track these creates a mess. It would be best to output all images to a temporary folder and to gitignore that folder.
- [x] `tenseg_video_strain_stress` and `tenseg_ex_force`, and perhaps others, have unused variables e.g. `pic_num`, and `a`. Please check and update codes accordingly.
- [x] It would be nice if the source for the manual was shared too so people can help edit that too.
