### 2022-04-13 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @diehlpk
@Muhao-Chen Please work on these comments soon.
