### 2022-04-16 b957183
Update tenseg_plot_strain_stress.m
