+++ Function_library/tenseg_plot_strain_stress.m
@@ -1,4 +1,4 @@
-function tenseg_video_strain_stress(data_out,ele_num,time)
+function tenseg_plot_strain_stress(data_out,ele_num,time)
 % /* This Source Code Form is subject to the terms of the Mozilla Public
 % * License, v. 2.0. If a copy of the MPL was not distributed with this
 % * file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -19,7 +19,7 @@ function tenseg_video_strain_stress(data_out,ele_num,time)
 color=['y','m','c','r','g','b','w','k'];
 name=['stress_strain3',data_out.material{1}];
 figure;
-%     set(gcf,'Position',get(0,'ScreenSize'));  %full screen
+% set(gcf,'Position',get(0,'ScreenSize'));  %full screen
 for i = 1:size(data_out.strain_t,2)
     if strcmp(data_out.material{1},'linear_elastic')|strcmp(data_out.material{1},'plastic')|strcmp(data_out.material{1},'multielastic')
         plot(reshape(data_out.strain_t(ele_num,:)',[],1),reshape(data_out.stress_t(ele_num,:)',[],1),'o','MarkerSize',5);
@@ -42,14 +42,13 @@ for i=1:numel(time)
         plot(data_out.strain_t(ele_num(j),pic(i)),data_out.stress_t(ele_num(j),pic(i)),'rs','LineWidth',2,'MarkerSize',18,'MarkerEdgeColor',color(j),'MarkerFaceColor',color(j));
         hold on;
     end
-
 end
 set(gcf,'color','w');
-%                 set(gca,'fontsize',18);
-set(gca,'fontsize', 18,'linewidth',1.15)
-set(gcf,'position',[100,100,400,350])
+% set(gca,'fontsize',18);
+set(gca,'fontsize', 18,'linewidth',1.15);
+set(gcf,'position',[100,100,400,350]);
 grid on;
-set(gca, 'XGrid','on'); % XÖáµÄÍø¸ñ
+set(gca, 'XGrid','on'); 
 ylabel('Stress (Pa)','fontsize',18);
 xlabel('Strain','fontsize',18);
 end
