### 2022-04-16 9e3d91d
Update tenseg_plot_strain_stress.m
