### 2022-04-17 3897fc1
delete unused variable: pic_num
