### 2022-04-17 d812317
delete unused variable a in 'tenseg_ex_force'
