+++ Software_Verification_and_Examples/Dynamics_Examples/03_double_pendulum/FEM_solution/main_double_pendulum_simple.m
@@ -92,7 +92,7 @@ out_dt=dt;            % output data interval(approximately, not exatly)
 out_tspan=interp1(tspan,tspan,0:out_dt:tf, 'nearest','extrap');  % output data time span
 
 % calculate external force and forced motion of nodes
-[w_t,dnb_t,dnb_d_t,dnb_dd_t,dz_a_t]=tenseg_ex_force(tspan,a,b,'step',gravity,[0;9.8;0],C,mass,3*3-1,0);
+[w_t,dnb_t,dnb_d_t,dnb_dd_t,dz_a_t]=tenseg_ex_force(tspan,b,'step',gravity,[0;9.8;0],C,mass,3*3-1,0);
 
 % give initial speed of free coordinates
 n0a_d=0*ones(numel(a),1);        %initial speed in X direction
