+++ Software_Verification_and_Examples/Dynamics_Examples/05_T_bar_Dynamics/main_Tbar_dynamics.m
@@ -121,7 +121,7 @@ end
 tspan=0:dt:tf;
 out_tspan=interp1(tspan,tspan,0:out_dt:tf, 'nearest','extrap');  % output data time span
 
-[w_t,dnb_t,dnb_d_t,dnb_dd_t,dz_a_t]=tenseg_ex_force(tspan,a,b,'step',gravity,[0;0;0],C,mass,3*3-1,0);
+[w_t,dnb_t,dnb_d_t,dnb_dd_t,dz_a_t]=tenseg_ex_force(tspan,b,'step',gravity,[0;0;0],C,mass,3*3-1,0);
 
 % give initial speed of free coordinates
 n0a_d=2*ones(numel(a),1);        %initial speed in X direction
