+++ Software_Verification_and_Examples/Dynamics_Examples/05_T_bar_Dynamics/tower.txt
@@ -4,7 +4,7 @@ finish
 /clear 
 /filename,tower  
 /title,the analysis of tower  
-!Unit:m£¬N£¬Pa£¬s
+!Unit:m，N，Pa，s
 
 /prep7
 !specify element type 
@@ -81,7 +81,7 @@ latt,1,,1,,,,3  !specify section area
 LMESH,ALL
 finish
 
-!First solve for self-equilibrium£©
+!First solve for self-equilibrium）
 /SOLU
 ANTYPE,0 
 NLGEO!consider large deformation 
