+++ Software_Verification_and_Examples/Statics_Examples/06_half_Tbar/main_half_Tbar_static.m
@@ -36,6 +36,7 @@ saveimg=0;              % save image or not (1) yes (0)no
 savedata=0;             % save data or not (1) yes (0)no
 savevideo=0;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
+savePath=fullfile(fileparts(mfilename('fullpath')),'data_temp'); %Save files in same folder as this code
 
 %% N C of the structure
 % Manually specify node positions (in 3D).
