+++ Software_Verification_and_Examples/Statics_Examples/04_Jansen_linkage/example1/main_Jansen_linkage_1.m
@@ -25,7 +25,7 @@ savedata=0;             % save data or not (1) yes (0)no
 savevideo=1;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
 tf = 1/substep;         % simulation time step
-savePath=fileparts(mfilename('fullpath')); %Save files in same folder as this code
+savePath=fullfile(fileparts(mfilename('fullpath')),'data_temp'); %Save files in same folder as this code
 
 %% N C of the structure
 % Manually specify node positions (accurate).
