+++ Software_Verification_and_Examples/Statics_Examples/02_3D_Dbar/main_folding_3d_Dbar.m
@@ -18,7 +18,7 @@
 % 10.Equilibrium calculation
 % 11. Plot and make video, output data to TECPLOT(optional)
 %%
-clc;clear;close all;
+clc;clear;close all; warning off;
 % Specify material properties
 [consti_data,Eb,Es,sigmab,sigmas,rho_b,rho_s]=material_lib('Steel_Q345','Steel_string');
 material{1}='linear_elastic'; % index for material properties:linear_elastic; multielastic; plastic
@@ -35,7 +35,7 @@ saveimg=0;              % save image or not (1) yes (0)no
 savedata=0;             % save data or not (1) yes (0)no
 savevideo=1;            % make video(1) or not(0)
 gravity=0;              % consider gravity 1 for yes, 0 for no
-savePath=fileparts(mfilename('fullpath')); %Save files in same folder as this code
+savePath=fullfile(fileparts(mfilename('fullpath')),'data_temp'); %Save files in same folder as this code
 
 %% N C of the structure
 % Manually specify node positions of a 3D D-Bar Unit.
