+++ Software_Verification_and_Examples/Dynamics_Examples/03_double_pendulum/Analytical_Solution/data_temp/README.md
@@ -1,0 +1,5 @@
+This is a local folder. 
+
+One can save images, data, and gifs to this temporary folder. 
+
+This folder will not be updated because of the gitignore command.
