### 2022-04-18 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Muhao-Chen
> * [x]  The function ansys_input_gp_00 still contains what looks like non-english symbols my system cannot parse. Please update
Thanks for your question. There is a Language option in the code, so one can choose to use English or Chinese. As you may see in ansys_input_gp_00 , line 27, language=0;   % 0 for English,1 for Chinese.

> * [x]  functions like `tenseg_mode` and `tenseg_video_strain_stress` save images to the current working directory. Change this throughout to be a local folder. Also using git to track these creates a mess. It would be best to output all images to a temporary folder and to gitignore that folder.
Thank you! We have followed your valuable suggestion and added the data_temp folder and used the gitignore.

> * [x]  `tenseg_video_strain_stress` and `tenseg_ex_force`, and perhaps others, have unused variables e.g. `pic_num`, and `a`. Please check and update codes accordingly.
Thank you for pointing this out. Yes, we have some unused variables, some of them are left because they may be used for different simulation cases and future development. For example, one wants to simulate an earthquake or impulse to the structure, so he/she can use the unused variables and modify the loading conditions a little bit. Thank you!

> * [x]  It would be nice if the source for the manual was shared too so people can help edit that too.
Thank you! We have uploaded the word source file.
