### 2022-04-18 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Muhao-Chen
@Kevin-Mattheus-Moerman @diehlpk Please have a check and let me know if you have any questions anytime! Thank you so much for your great help in improving the codes!
