### 2022-04-21 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @diehlpk
@Kevin-Mattheus-Moerman Could you please have a look, I would like to finish the paper soon.
