### 2022-04-21 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
@Muhao-Chen

- [x] The function `tenseg_video_strain_stress` still saves a file locally (outside of data_temp folder). If you made `name` an input to the function then the codes calling this function can specify the save path to point to `data_temp`.
- [x] Your `.gitignore` file is not working properly, you could change it to be like below. Note the change relating to the `data_temp` folders:
```
# Ignore MATLAB temporary files
*.asv

#Ignore temp file for running tests
*.mat

# Ignore gif animations files
*.gif

# Ignore png images
*.png

# Ignore txt files
*.txt

# Ignore folder content
**/data_temp
```
Specifically I removed the comment symbol `#` and added ** so that any data_temp folder is ignore (you are using multiple). Note that you are currently ignoring all txt, png, and gif files. This also means that changes in the Videos folder and other images relating to documentation may not be tracked. You may consider perhaps if you want this or if these are no longer needed after ignoring those data_temp files.

- [x] Recommendation: The Jansen codes contain `machanism`, which should read `mechanism`
- [x] Recommendation: The code `main_Double_Pendulum_Analytical.m` uses uppercase letters while `main_double_pendulum_simple` uses lowercase letters. I recommend lower case throughout (or to at least have them be the same).
- [x] Unfortunately it is not very repeatable but I now get this type of error every now and then when running the tests:

```
Error using matlab.internal.imagesci.wgifc
Can only append to GIF89a format GIFs.

Error in writegif (line 306)
matlab.internal.imagesci.wgifc(mat, map, filename,writemode,disposalmethod,delaytime,...

Error in imwrite (line 566)
        feval(fmt_s.write, data, map, filename, paramPairs{:});

Error in tenseg_savegif_forever (line 56)
    imwrite(f,map,filename,'writemode','append','delaytime',dt);

Error in tenseg_video_slack (line 45)
        tenseg_savegif_forever(name);

Error in main_Jansen_linkage_1 (line 152)
tenseg_video_slack(n_t,C_b,C_s,l0_t,index_s,[],[0,90],[-80,80,-85,50,0,70],min(substep,30),name,savevideo,material{2})
```
