+++ Software_Verification_and_Examples/auto_statics_test.m
@@ -27,7 +27,7 @@ testFileList={'main_tower_static.m',...
               'main_Jansen_linkage_2.m',...
               'main_Jansen_linkage_3.m',...
               'main_cable_dome.m',...
-              'main_half_Tbar_static.m'...
+              'main_half_tbar_static.m'...
               };
 testFolderList={'01_tower',...
                 '02_3D_Dbar',...
