### 2022-04-21 829488f
Update auto_dynamics_test.m
