+++ Software_Verification_and_Examples/auto_dynamics_test.m
@@ -22,10 +22,10 @@ clear; close all; clc;
 originFolder=fileparts(mfilename('fullpath'));
 testFileList={'main_tower_dynamic_simple.m',...
               'main_10seg_truss_simple.m',...
-              'main_Double_Pendulum_Analytical.m',...
+              'main_double_pendulum_analytical.m',...
               'main_double_pendulum_simple.m',...
               'main_lander.m',...
-              'main_Tbar_dynamics'...
+              'main_tbar_dynamics'...
               };
 testFolderList={'01_tower',...
                 '02_cantilever_truss',...
