+++ Software_Verification_and_Examples/auto_dynamics_test.m
@@ -24,8 +24,8 @@ testFileList={'main_tower_dynamic_simple.m',...
               'main_10seg_truss_simple.m',...
               'main_double_pendulum_analytical.m',...
               'main_double_pendulum_simple.m',...
-              'main_Lander.m',...
-              'main_T_bar_dynamics'...
+              'main_lander.m',...
+              'main_tbar_dynamics'...
               };
 testFolderList={'01_tower',...
                 '02_cantilever_truss',...
