+++ Software_Verification_and_Examples/test_all.m
@@ -1,4 +1,18 @@
-function test_all
-
+function test_all()
+%% Automated test 
+% By runing this code, user can see all the examples given in this software. 
+% This automated test code was derived from codes from the GIBBON project.
+% The source code is here: 
+% https://github.com/gibbonCode/GIBBON/blob/master/lib/testGibbon.m
+% The license for GIBBON is given below. 
+% License: <https://github.com/gibbonCode/GIBBON/blob/master/LICENSE>
+% And the copyright information is:
+% Copyright (C) 2006-2021 Kevin Mattheus Moerman and the GIBBON
+% contributors.
+% You should have received a copy of the GNU General Public License
+% along with this program.  If not, see <http://www.gnu.org/licenses/>.
+% Please include their licence and Copyright information when you use this
+% file. 
+%% 
 auto_statics_test;
 auto_dynamics_test;
