### 2022-04-21 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Muhao-Chen
@Kevin-Mattheus-Moerman Thanks for the comments. All the issues have been fixed. Please have a check, and thank you again!
