### 2022-04-22 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Kevin-Mattheus-Moerman
@Muhao-Chen thanks for making those changes. That last error :point_up: still occurs some times but again it is not repeatable, could be a linux graphics issue, I am not sure, often if I rerun the code it doesn't happen so it is odd. One last issue, when running the dynamics tests using `auto_dynamics_test`, I got an error as MATLAB could not find `main_tbar_dynamics`. This code is currently called `main_Tbar_dynamics` (note the capital T). To fix it you need to rename it, or you have to change the code in `auto_dynamics_test` to have a capital T.

If you fix this last point I believe all my issues are addressed.
