### 2022-04-22 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @Muhao-Chen
@Kevin-Mattheus-Moerman Thanks for your help! Yeah, probably, the error does show up on my windows environment. I have renamed main_Tbar_dynamics.m as main_tbar_dynamics.m. Thank you so much!

@diehlpk Please also have a check on everything. Thank you for your time and effort!
