### 2022-04-22 Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM#3 @diehlpk
@Muhao-Chen I will have a check once @Kevin-Mattheus-Moerman checked all boxes in the main review thread.
