### 2022-05-10 #3390 @Kevin-Mattheus-Moerman
@Muhao-Chen some minor comments on the paper

- [ ] Check if `prestress` should be written as `pre-stress`
- [ ] `expansive` should be `expensive`
- [ ] Recomendation: consider expanding the various sets of listed items in the text on new lines e.g.

1. First listed item.
2. Second listed item.

@diehlpk I am nearly done with my review. I went through the paper and provided some feedback above.

I've one last point remaining for my review, and that is to test the link with ANSYS. I should have that done by the next couple of days.
