+++ JOSS_paper/paper.md
@@ -19,7 +19,7 @@ authors:
     orcid: 0000-0001-6503-9115
     affiliation: 2
 affiliations:
- - name: College of Civil Engineering, Zhejiang University of Technology, Hanzhou, Zhejiang, China
+ - name: College of Civil Engineering, Zhejiang University of Technology, Hangzhou, Zhejiang, China
    index: 1
  - name: Department of Aerospace Engineering, Texas A&M University, College Station, Texas, USA
    index: 2
@@ -29,24 +29,44 @@ bibliography: paper.bib
 
 # Summary
 
-The name of this software, TsgFEM, is suggested to be pronounced as Tenseg FEM. The purpose of this package is to facilitate the analysis of statics and dynamics of any tensegrity structures. The software allows one to perform topology design, statics and dynamic FEM simulation of any tensegrity systems. 
+The name of this software, TsgFEM, is suggested to be pronounced as Tenseg FEM. The purpose of this package is to facilitate the analysis of statics and dynamics of any tensegrity structures. The software allows one to perform topology design, statics, and dynamic FEM simulation of any tensegrity systems. 
 
-This software allows one to do the following studies, but not limited to the listed items. Topology design: 1). Modeling any tensegrity structures by nodal coordinates and the node's connectivity information. 2). Specifying the constraints of nodal coordinates, and grouping structure members and force them to have same force densities. Statics: 1). Conducting structure equilibrium configuration, prestress design, and stiffness studies. 2). Performing prestress and mechanism modes analysis. 3). Checking stiffness, stability, and robustness in terms of prestress, materials, and geometric information of the structure. 4). Conducting studies on form-finding of tensegrity systems. 5). Simulating the forced motion of structures. 6) Studying the feasibility of pseudo-static deployment trajectories. Dynamics: 1). Rigid body dynamics with acceptable errors. This is achieved by setting relatively high stiffness for bars in the dynamics simulation. 2). FEM dynamics simulation with elastic or plastic deformations in the presence of various kinds of boundary conditions, such as fix any nodes in any direction, apply static or dynamic external forces (i.e., gravitational force, some specified forces, or arbitrary seismic vibrations). 3). Modal analysis, including natural frequency and corresponding modes. 4). An interface towards structural control by a compact state-space form of a linear model.
+This software allows one to do the following studies but is not limited to the listed items. 
+Topology design: 
+1. Modeling any tensegrity structures by nodal coordinates and the node's connectivity information. 
+2. Specifying the constraints of nodal coordinates and grouping structure members and forcing them to have the same force densities. 
+Statics: 
+1. Conducting structure equilibrium configuration, pre-stress design, and stiffness studies. 
+2. Performing pre-stress and mechanism modes analysis. 
+3. Checking stiffness, stability, and robustness in terms of pre-stress, materials, and geometric information of the structure. 
+4. Conducting studies on form-finding of tensegrity systems. 
+5. Simulating the forced motion of structures. 
+6. Studying the feasibility of pseudo-static deployment trajectories. 
+Dynamics: 
+1. Rigid body dynamics with acceptable errors. This is achieved by setting relatively high stiffness for bars in the dynamics simulation. 
+2. FEM dynamics simulation with elastic or plastic deformations in the presence of various kinds of boundary conditions, such as fixing any nodes in any direction, applying static or dynamic external forces (i.e., gravitational force, some specified forces, or arbitrary seismic vibrations). 
+3. Modal analysis, including natural frequency and corresponding modes. 
+4. An interface towards structural control by a compact state-space form of a linear model.
 
 
[…]
