+++ JOSS_paper/paper.md
@@ -68,9 +68,9 @@ Based on the following assumptions, the FEM equations for the statics and dynami
 
 # Statement of need
 
-For the FEM analysis for tensegrity structures, little research has been conducted. For example, Zogoul et al. conducted a static study for a tensegrity bridge by FEM [@zgoul2012static]. Jensen et al. showed the finite element analysis of tensegrity structures in offshore aquaculture installations by ABAQUS [@jensen2007finite]. Kan et al. formulated the dynamics of clustered tensegrity structures by FEM [@kan2018nonlinear]. However, most of these analyses are either using commercial software or assuming the structure members are elastic. Few software packages have been developed for the analysis of tensegrity statics and dynamics. For example, STEDY [@tadiparthi2019stedy] is a package for conducting tensegrity dynamics with rigid bars based on the Lagrangian method. MOTES [@goyal2019motes] is software for the analysis of both statics and dynamics with rigid bars and linear elastic strings. Both of them are rigid body dynamics and developed in non-minimum Cartesian coordinates. However, for many applications, the elastic or plastic deformation of structure members cannot be neglected. It is possible to do the FEM analysis by commercial software, i.e., ANSYS and ABAQUS, but the licenses are expensive, and the modeling and setups require much experience. To this end, we derive a closed-form dynamics equation based on Lagrangian's method with a nodal vector as the generalized coordinate, allowing the user to perform simulations for bars and strings with any given strain-stress properties. The proposed dynamics equation is in a compact form.
+For the FEM analysis for tensegrity structures, little research has been conducted. For example, Zogoul et al. conducted a static study for a tensegrity bridge by FEM [@zgoul2012static]. Jensen et al. showed the finite element analysis of tensegrity structures in offshore aquaculture installations by ABAQUS [@jensen2007finite]. Kan et al. formulated the dynamics of clustered tensegrity structures by FEM [@kan2018nonlinear]. However, most of these analyses use commercial software or assume the structure members are elastic. Few software packages have been developed for the analysis of tensegrity statics and dynamics. For example, STEDY [@tadiparthi2019stedy] is a package for conducting tensegrity dynamics with rigid bars based on the Lagrangian method. MOTES [@goyal2019motes] is software for the analysis of both statics and dynamics with rigid bars and linear elastic strings. Both of them are rigid body dynamics and developed in non-minimum Cartesian coordinates. However, for many applications, the elastic or plastic deformation of structure members cannot be neglected. It is possible to do the FEM analysis by commercial software, i.e., ANSYS and ABAQUS, but the licenses are expensive, and the modeling and setups require much experience. To this end, we derive a closed-form dynamics equation based on Lagrangian's method with a nodal vector as the generalized coordinate, allowing the user to perform simulations for bars and strings with any given strain-stress properties. The proposed dynamics equation is in a compact form.
 
-The developed software is capable of dealing with large deformation for structures with elastic or plastic materials. The results of the software are compared with analytical solutions as well as commercial software ANSYS. In fact, this software also provides an interface to ANSYS. That is, one can modify the settings in the example source codes based on the simulation needs and run the codes. This software can generate a text file that allows users to run the simulation in ANSYS automatically by using the APDL (Ansys Parametric Design Language) interface. The software has been used for the design and analysis of various tensegrity structures, i.e., tensegrity lunar tower, tensegrity morphing wings, tensegrity cable domes, tensegrity lander, tensegrity dolphin, where we integrate structure, control design, and signal processing to get the required performance and control law.
[…]
