### 2022-05-16 #3390 @diehlpk
> @diehlpk Thanks for your question! Actually, in the research body (papers from a large number of journals), both words are used to describe the same thing. So, I believe people in the structure field can understand both. Thank you!

Ok, so pick and check the checkbove.
