### 2022-06-24 #3390 @Kevin-Mattheus-Moerman
@diehlpk I have checked with several colleagues on evaluating the ANSYS functionality but have so far not been able to get any to commit to helping out to check this aspect of the software functionality.

Can @HaoZeke and @likask comment if they evaluated this functionality?

I have however very thoroughly tested the MATLAB aspects of the software. Despite not being able to verify the ANSYS link, I have ticked that final "functionality" box above :point_up:, and do recommend publication of this work in JOSS, since the ANSYS link is a very minor component of this overall software package and its functionality.
