### 2022-06-25 #3390 @diehlpk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.2307/3103256 is OK
> - 10.1177/026635119601-204 is OK
> - 10.1016/j.compstruct.2020.112454 is OK
> - 10.1073/pnas.141199598 is OK
> - 10.1038/nnano.2010.107 is OK
> - 10.1007/978-0-387-74242-7 is OK
> - 10.1016/j.compstruct.2020.112188 is OK
> - 10.1016/j.engstruct.2009.12.042 is OK
> - 10.1016/j.paerosci.2020.100692 is OK
> - 10.1061/9780784483374.100 is OK
> - 10.1115/1.4037565 is OK
> - 10.1109/ICRA.2015.7139590 is OK
> - 10.1089/soro.2019.0142 is OK
> - 10.1109/TCST.2020.2975138 is OK
> - 10.1016/j.mechrescom.2020.103480 is OK
> - 10.1089/soro.2018.0079 is OK
> - 10.1088/1742-6596/1885/5/052024 is OK
> - 10.1016/j.aquaeng.2007.01.001 is OK
> - 10.1016/j.compstruct.2017.12.050 is OK
> - 10.21105/joss.01042 is OK
> - 10.21105/joss.01613 is OK
>
> MISSING DOIs
>
> - 10.1016/j.compstruct.2021.114838 may be a valid DOI for title: Tensegrity system dynamics based on finite element method
>
> INVALID DOIs
>
> - None
> ```

@Muhao-Chen Please add the missing DOI to the paper before you do the release.
