+++ JOSS_paper/paper.bib
@@ -174,7 +174,8 @@
   volume={280},
   pages={114838},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.compstruct.2021.114838}
 }
 @inproceedings{zgoul2012static,
   title={Static Analysis of a Tensegrity Bridge using the Finite Element Method},
