### 2022-06-27 #3390 @diehlpk
@Muhao-Chen that above is the archive. I was speaking about a new release.

https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/releases

The current version 1.1 is from Jun 09, 2021. Please generate a new release with all the JOSS changes. Depending on the changes, you could consider a minor release 1.1.1 or a major release 1.2.
