### 2022-06-27 #3390 @Muhao-Chen
@diehlpk Thank you so much for the clarification. I have released the new version v1.2.
https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/releases
Please let me know if you have any questions. Thank you!!
