### 2022-06-28 #3390 @arfon
@Muhao-Chen – I've made a couple of minor fixes to your paper formatting here: https://github.com/Muhao-Chen/Tensegrity_Finite_Element_Method_TsgFEM/pull/7
